% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rda.R
\name{calc_kqs}
\alias{calc_kqs}
\title{Calculate key quantities}
\usage{
calc_kqs(
  returned_kq,
  prepared_data = NULL,
  dat = NULL,
  final = NULL,
  relative = NULL,
  module = NULL,
  ...
)
}
\arguments{
\item{returned_kq}{String. Key quantity to be returned.}

\item{prepared_data}{Data frame. The prepared_data object.}

\item{dat}{Data frame. The original data object.}

\item{final}{ggplot2 object. The final figure.}

\item{relative}{Logical. TRUE/FALSE; specify whether to set y-axis values relative to the ref_line value.}

\item{module}{String. Module name found in `dat`.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
The value of the specified key quantity.
}
\description{
Calculate key quantities
}
\examples{
\dontrun{
calc_kqs(
  returned_kq = "F.min",
  dat = dat
)
}
}
