% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{create_latex_table}
\alias{create_latex_table}
\title{Create the rda package for a plot or table}
\usage{
create_latex_table(data, caption, label)
}
\arguments{
\item{data}{A dataframe-based table}

\item{caption}{A string comprising the table caption}

\item{label}{A string comprising the table label}
}
\value{
A table based in LaTeX.
}
\description{
Create the rda package for a plot or table
}
\examples{
create_latex_table(
  data = as.data.frame(head(mtcars, 6)),
  caption = "My caption",
  label = "My label"
)
}
