% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rda.R
\name{export_kqs}
\alias{export_kqs}
\title{Export updated key quantities template}
\usage{
export_kqs(...)
}
\arguments{
\item{...}{Key quantities whose values should be added to the
exported "key_quantities.csv"}
}
\value{
Exports a file based on key quantities template, with values
added next to the names of the key quantities specified as ellipsis
arguments. File is saved as "key_quantities.csv" to the working directory.
}
\description{
Export updated key quantities template
}
\examples{
\dontrun{
export_kqs(
  F.min,
  F.max
)
}
}
