% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rda.R
\name{extract_caps_alttext}
\alias{extract_caps_alttext}
\title{Extract captions and alternative texts}
\usage{
extract_caps_alttext(topic_label = NULL, fig_or_table = NULL, dir = getwd())
}
\arguments{
\item{topic_label}{String. Figure or table label. These
labels are found in the "label" column of the "captions_alt_text.csv" file
and are used to link the figure or table with its caption/alt text.

Default: NULL}

\item{fig_or_table}{String. Indicates whether the plot is a figure or table.

Default: NULL}

\item{dir}{Path. The directory containing the "captions_alt_text.csv" file.

Default: the working directory (`getwd()`)}
}
\value{
A figure's caption and alternative text, in a list, or a table's caption.
}
\description{
Extract a figure or table's caption and alternative text for usage when
generating a figure or table. Typically used before stockplotr::export_rda().
}
\examples{
\dontrun{
extract_caps_alttext(
  topic_label = "biomass",
  fig_or_table = "figure",
  dir = here::here()
)

extract_caps_alttext(
  topic_label = "biomass_abundance_catch",
  fig_or_table = "table",
  dir = getwd()
)
}
}
\seealso{
[export_rda()], [create_rda()]
}
