% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rda.R
\name{fill_in_kqs}
\alias{fill_in_kqs}
\title{Substitute key quantities' values into template}
\usage{
fill_in_kqs(df, ...)
}
\arguments{
\item{df}{Dataframe created by importing "key_quantity_template.csv" or
"key_quantities.csv", empty or partially-filled templates with key
quantity names and other associated information}

\item{...}{Key quantity objects whose values will be added to the output
dataframe}
}
\value{
Dataframe based on key quantities template that contains 
newly-added values of key quantities indicated via ellipsis. The dataframe
is input for [fill_in_kqs()].
}
\description{
Substitute key quantities' values into template
}
\examples{
\dontrun{
fill_in_kqs(
  df = data,
  F.min,
  F.max)
}
}
