% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rda.R
\name{insert_kqs}
\alias{insert_kqs}
\title{Insert key quantities into the captions and alternative text file}
\usage{
insert_kqs(...)
}
\arguments{
\item{...}{Key quantities whose values should be added to the
exported "captions_alt_text.csv"}
}
\value{
Exports a file ("captions_alt_text.csv") containing captions 
and alternative text for figures and tables, with key quantities inserted
into the "captions_alt_text_template.csv" template's placeholders.
}
\description{
Insert key quantities into the captions and alternative text file
}
\examples{
\dontrun{
insert_kqs(
  F.min,
  F.max)
}
}
