% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{merge_error}
\alias{merge_error}
\title{Rename columns and merge estimate and uncertainty columns for table presentation}
\usage{
merge_error(table_data, uncert_lab, fleets, label, unit_label)
}
\arguments{
\item{table_data}{list of dataframes that will be eventually turned into tables}

\item{uncert_lab}{uncertainty label. Typically inherited from another function
but is the exact string of the uncertainty in the data (e.g., "sd", "se", "cv",
"uncertainty").)}

\item{fleets}{Vector of fleet names.}

\item{label}{Label name of target quantity that is being presented by the table.}

\item{unit_label}{String. The units of the estimate being presented in the table.}
}
\value{
List of formatted dataframes that contain column names formatted
for a table along with a merge of values in the estimate and error columns
to reduce redundancy in the table.
}
\description{
Rename columns and merge estimate and uncertainty columns for table presentation
}
