% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discard.R
\name{plot_discard}
\alias{plot_discard}
\title{Plot discards}
\usage{
plot_discard(
  dat,
  group = NULL,
  facet = NULL,
  unit_label = "mt",
  era = NULL,
  lbs = FALSE,
  scale_amount = 1,
  interactive = TRUE,
  module = NULL,
  make_rda = FALSE,
  figures_dir = getwd()
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{group}{String. Single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{Character vector. Column name or names used for faceting.

Default: NULL}

\item{unit_label}{String. Spawning biomass unit as indicated from the data. 
When setting a unit of pounds please use one of the following keywords: 
"lbs", "pounds", "lb". For example "1000s of lbs" would still be recognized 
as pounds. Do not set unit label to the desired order of magnitude rather the
unit label reflected in the data. Please refer to the argument `scale_amount`.

Default: NULL}

\item{era}{String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{lbs}{`r lifecycle::badge('deprecated')` Logical. TRUE/FALSE; indicate whether to convert the y-axis values from
kilograms to pounds. The default units match the default in the
unit_label argument - 'mt'. -- Deprecate

Default: `FALSE`}

\item{scale_amount}{Number. A number to scale the y-axis values. If > 1, the
 magnitude of the scale amount will be added to the unit label 
 (e.g. scale_amount = 1000 with a unit label of "1000s of lbs" will result
  in "thousands of 1000s of lbs")

Default: 1}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{module}{Character vector. (Optional) Module name found in `dat$module_name`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{Path. Path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}
}
\value{
A ggplot showing discards over time, usually by fleet.
}
\description{
Plot discards as a line graph.
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a `ggplot2` object or export an .rda object
containing associated caption and alternative text for the figure.
}
\note{
All plotting functions automatically recognize indexing variables and will
use them in groupings and/or facetting. @seealso [process_data()].
}
\examples{
plot_discard(
  dat = stockplotr:::example_data,
  module = "DISCARD_OUTPUT"
)
}
\seealso{
[convert_output()], [plot_timeseries()], [calculate_reference_point()], [reference_line()], [filter_data()], [process_data()], [export_kqs()], [insert_kqs()], [create_rda()]
}
