% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_error}
\alias{plot_error}
\title{Create plot with error}
\usage{
plot_error(
  dat,
  x = "year",
  y = "estimate",
  geom = "point",
  group = NULL,
  facet = NULL,
  xlab = "Year",
  ylab = NULL,
  hline = TRUE,
  ...
)
}
\arguments{
\item{dat}{filtered data frame from standard output file(s) preformatted for
the target label from \link[stockplotr]{filter_data}}

\item{x}{a string of the column name of data used to plot on the x-axis

Default: "year"}

\item{y}{a string of the column name of data used to plot on the y-axis

Default: "estimate"}

\item{geom}{type of geom to use for plotting found in ggplot2 (e.g. "point",
"line", etc.).

Default: "line"

Options: "point" and "area"}

\item{group}{a string of a single column that groups the data.
Currently can only have one level of grouping.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.)

Default: NULL}

\item{xlab}{a string of the x-axis label

Default: "Year"}

\item{ylab}{a string of the y-axis label. If NULL, it will be set to the name
 of `y`.

Default: NULL}

\item{hline}{indicate true or false to place a horizontal line at 1

Default: `TRUE`}

\item{...}{inherited arguments from internal functions from ggplot2::geom_xx}
}
\value{
Create a plot with error for a stock assessment report.
}
\description{
Create plot with error
}
\details{
The user can create a line, point, or area plot, where the x-axis is
year and y can vary for any time series quantity. Currently, grouping is
restricted to one group where faceting can be any number of facets.
}
