% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landings.R
\name{plot_landings}
\alias{plot_landings}
\title{Plot observed landings}
\usage{
plot_landings(
  dat,
  unit_label = "mt",
  geom = "line",
  group = NULL,
  facet = NULL,
  lbs = FALSE,
  era = NULL,
  scale_amount = 1,
  module = NULL,
  interactive = TRUE,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{String. Spawning biomass unit as indicated from the data. 
When setting a unit of pounds please use one of the following keywords: 
"lbs", "pounds", "lb". For example "1000s of lbs" would still be recognized 
as pounds. Do not set unit label to the desired order of magnitude rather the
unit label reflected in the data. Please refer to the argument `scale_amount`.

Default: NULL}

\item{geom}{String. Geom used for the plot.

Default: "line".

Options: "line", "point", or "area"}

\item{group}{String. Single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{Character vector. Column name or names used for faceting.

Default: NULL}

\item{lbs}{`r lifecycle::badge('deprecated')` Logical. TRUE/FALSE; indicate whether to convert the y-axis values from
kilograms to pounds. The default units match the default in the
unit_label argument - 'mt'. -- Deprecate

Default: `FALSE`}

\item{era}{String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{scale_amount}{Number. A number to scale the y-axis values. If > 1, the
 magnitude of the scale amount will be added to the unit label 
 (e.g. scale_amount = 1000 with a unit label of "1000s of lbs" will result
  in "thousands of 1000s of lbs")

Default: 1}

\item{module}{Character vector. (Optional) Module name found in `dat$module_name`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{Path. Path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}

\item{...}{Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
A plot showing cumulative landings over time.
}
\description{
Plot observed landings
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a `ggplot2` object or export an .rda object
containing associated caption and alternative text for the figure.
}
\examples{
plot_landings(
  dat = stockplotr:::example_data,
  unit_label = "metric tons",
  group = "fleet",
  interactive = FALSE,
  make_rda = FALSE
)
plot_landings(
  dat = stockplotr:::example_data,
  unit_label = "metric tons",
  facet = "fleet",
  interactive = FALSE,
  make_rda = FALSE
)
}
\seealso{
[convert_output()], [filter_data()], [process_data()], [plot_timeseries()], [export_kqs()], [insert_kqs()], [create_rda()]
}
