% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mortality_at_age.R
\name{plot_mortality_at_age}
\alias{plot_mortality_at_age}
\title{Plot natural mortality (M) at age}
\usage{
plot_mortality_at_age(dat, make_rda = FALSE, rda_dir = getwd())
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the .rda will be exported to the folder indicated in the argument "rda_dir".
Default is FALSE.}

\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Plot natural mortality at age from a stock assessment model as found in a NOAA
stock assessment report.
}
\description{
Plot natural mortality (M) at age
}
