% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment_deviations.R
\name{plot_recruitment_deviations}
\alias{plot_recruitment_deviations}
\title{Plot recruitment deviations}
\usage{
plot_recruitment_deviations(
  dat,
  module = NULL,
  era = NULL,
  interactive = TRUE,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{module}{Character vector. (Optional) Module name found in `dat$module_name`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{era}{String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{Path. Path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}

\item{...}{Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
A plot showing recruitment deviations relative to one, over time.
}
\description{
Plot recruitment deviations
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a `ggplot2` object or export an .rda object
containing associated caption and alternative text for the figure.
}
\note{
All plotting functions automatically recognize indexing variables and will
use them in groupings and/or facetting.
}
\examples{
plot_recruitment_deviations(
  dat = stockplotr:::example_data,
  interactive = FALSE,
  size = 1.5
)
plot_recruitment_deviations(
  dat = stockplotr:::example_data,
  era = "early",
  interactive = FALSE,
  shape = 2
)
}
\seealso{
[convert_output()], [plot_error()], [filter_data()],
[process_data()], [export_kqs()], [insert_kqs()], [create_rda()]
}
