% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawning_biomass.R
\name{plot_spawning_biomass}
\alias{plot_spawning_biomass}
\title{Plot spawning biomass (SB)}
\usage{
plot_spawning_biomass(
  dat,
  geom = "line",
  group = NULL,
  facet = NULL,
  ref_line = "msy",
  unit_label = NULL,
  era = NULL,
  lbs = lifecycle::deprecate(),
  module = NULL,
  scale_amount = 1,
  relative = FALSE,
  make_rda = FALSE,
  figures_dir = getwd(),
  interactive = TRUE,
  ...
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{geom}{String. Geom used for the plot.

Default: "line".

Options: "line", "point", or "area"}

\item{group}{String. Single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{Character vector. Column name or names used for faceting.

Default: NULL}

\item{ref_line}{String or named vector. Reference point name (and value if
custom). When multiple models are being compared, a reference line for each
model will appear, if desired. The following options can be used to customize
the reference line:
1. all lines are comparing same ref_line = "msy", ref_line = "target", ref_line = "unfished", etc
2. custom input vector - ref_line = c("sable23_msy"=30, "sable25_msy"=40) -- user must indicate which model in label otherwise it will assign in order of dat
3. input vector of labels - ref_line = c("msy", "target")

Default: "msy"

Options: (including, but not limited to) "target", "msy", and "unfished"
If the reference point is not found in the data, set ref_line = c("name" = value).}

\item{unit_label}{String. Spawning biomass unit as indicated from the data. 
When setting a unit of pounds please use one of the following keywords: 
"lbs", "pounds", "lb". For example "1000s of lbs" would still be recognized 
as pounds. Do not set unit label to the desired order of magnitude rather the
unit label reflected in the data. Please refer to the argument `scale_amount`.

Default: NULL}

\item{era}{String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{lbs}{`r lifecycle::badge('deprecated')` Logical. TRUE/FALSE; indicate whether to convert the y-axis values from
kilograms to pounds. The default units match the default in the
unit_label argument - 'mt'. -- Deprecate

Default: `FALSE`}

\item{module}{Character vector. (Optional) Module name found in `dat$module_name`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{scale_amount}{Number. A number to scale the y-axis values. If > 1, the
 magnitude of the scale amount will be added to the unit label 
 (e.g. scale_amount = 1000 with a unit label of "1000s of lbs" will result
  in "thousands of 1000s of lbs")

Default: 1}

\item{relative}{Logical. TRUE/FALSE; specify whether to set y-axis values relative to
the ref_line value.

Default: `FALSE`}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{Path. Path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{...}{Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
A plot showing spawning biomass over time.
}
\description{
Plot spawning biomass with a reference line as a dashed line. The figure can
also be made relative to this reference line rather than in absolute units.
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a `ggplot2` object or export an .rda object
containing associated caption and alternative text for the figure.
}
\note{
All plotting functions automatically recognize indexing variables and will
use them in groupings and/or facetting. @seealso [process_data()].
}
\examples{
plot_spawning_biomass(
  dat = stockplotr:::example_data,
  geom = "line",
  ref_line = "msy",
  unit_label = "mt",
  scale_amount = 1000,
  interactive = FALSE,
  module = "TIME_SERIES",
  linewidth = 1.5
)
plot_spawning_biomass(
  dat = stockplotr:::example_data,
  ref_line = c("target" = 10),
  interactive = FALSE,
  module = "TIME_SERIES"
)
\dontrun{
plot_spawning_biomass(
  dat = list("2023" = model1, "2025" = model2),
  ref_line = c("SBF30" = 1500, "SBF40" = 2300)
)
}
}
\seealso{
[convert_output()], [plot_timeseries()], [calculate_reference_point()], [reference_line()], [filter_data()], [process_data()], [export_kqs()], [insert_kqs()], [create_rda()]
}
