% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series trends}
\usage{
plot_timeseries(
  dat,
  x = "year",
  y = "estimate",
  geom = "line",
  xlab = "Year",
  ylab = NULL,
  group = NULL,
  facet = NULL,
  ...
)
}
\arguments{
\item{dat}{data frame. filtered data frame from standard output file(s) preformatted for
the target label from \link[stockplotr]{filter_data}}

\item{x}{string. a string of the column name of data used to plot on the x-axis

Default: "year"}

\item{y}{string. a string of the column name of data used to plot on the y-axis

Default: "estimate"}

\item{geom}{string. type of geom to use for plotting found in ggplot2 (e.g. "point",
"line", etc.).

Default: "line"

Options: "point" and "area"}

\item{xlab}{string. a string of the x-axis label

Default: "Year"}

\item{ylab}{string. a string of the y-axis label. If NULL, it will be set to the name
 of `y`.

Default: NULL}

\item{group}{string. a string of a single column that groups the data.
Currently can only have one level of grouping.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{character vector. a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.)

Default: NULL}

\item{...}{dots. inherited arguments from internal functions from ggplot2::geom_xx}
}
\value{
Create a time series plot for a stock assessment report.
}
\description{
Plot time series trends
}
\details{
The user can create a line, point, or area plot, where the x-axis is
year and y can vary for any time series quantity. Currently, grouping is
restricted to one group where faceting can be any number of facets.
}
\examples{
\dontrun{
plot_timeseries(dat,
  x = "year",
  y = "estimate",
  geom = "line",
  xlab = "Year",
  ylab = "Biomass",
  group = "fleet",
  facet = "area"
)
}
}
