% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Processing for figures}
\usage{
process_data(dat, group = NULL, facet = NULL, lbs = FALSE)
}
\arguments{
\item{dat}{Data frame. Pre-filtered data from \link[stockplotr]{filter_data} following a
long format data.}

\item{group}{String. Indexing variable of the data.
To remove any groups and summarize data across all factors, set
group = "none". Setting group = NULL will keep any grouping
variables identified in the data by the function.

Default: NULL

Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{Character vector. Faceting variable(s) of the data.

Default: NULL}

\item{lbs}{Logical. TRUE/FALSE; indicate whether to convert the y-axis values from
kilograms to pounds. The default units match the default in the
unit_label argument - 'metric tons'.

Default: `FALSE`}

\item{method}{String. Method for summarizing data when group
is set to "none".

Default: "sum"

Options: "sum" or "mean"}
}
\value{
List of 3 objects:
\item{data}{A data frame of the processed data ready for plotting.}
\item{group}{Grouping variable of the data.
If NULL, no grouping variable is identified. If not NULL, the function
will verify that the data is indexed by this variables, otherwise it will
overwrite it to a different, valid indexed variable or NULL.}
\item{facet}{Faceting variable(s) identifying the
variable(s) of the data. If NULL, no faceting variable is identified. Any
identified indexed variables found in this function will be added to facet.}
}
\description{
Processing for figures
}
\details{
Automatically detects potential grouping and faceting data from a
dataframe output from \link[stockplotr]{filter_data}.
}
\examples{
{
  filtered <- stockplotr::filter_data(
    dat = stockplotr:::example_data,
    label_name = "fishing_mortality$",
    geom = "line",
    module = "TIME_SERIES"
  )
  process_data(dat = filtered, method = "sum")
}
}
