% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_table}
\alias{process_table}
\title{Processing for tables}
\usage{
process_table(dat, group = NULL, method = "sum", label = NULL)
}
\arguments{
\item{dat}{Pre-filtered data from \link[stockplotr]{filter_data} following a
long format data.}

\item{group}{A string identifying the indexing variable of the data.
To remove any groups and summarize data across all factors, set
group = "none". Setting group = NULL will keep any grouping
variables identified in the data by the function.

Default: NULL

Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{method}{A string describing the method of summarizing data when group
is set to "none".

Default: "sum"

Options: "sum" or "mean"}

\item{label}{A string or vector of strings identifying the label values to filter the data.

Default: NULL}
}
\value{
A dataframe of processed data ready for formatting into a table.
}
\description{
Processing for tables
}
\details{
Input is an object created with \link[stockplotr]{filter_data}.
}
\examples{
{
  filtered <- filter_data(
    dat = stockplotr:::example_data,
    label = "landings",
    geom = "line",
    era = "time"
  )
  process_table(dat = filtered, method = "sum")
}
}
