% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{reference_line}
\alias{reference_line}
\title{Preformatted reference line}
\usage{
reference_line(label_name, ref_line, scale_amount = 1, model_name = "1")
}
\arguments{
\item{label_name}{String. Name of the quantity that users want to
extract the reference point from}

\item{ref_line}{String. Reference point(s)}

\item{scale_amount}{Number. A number to scale the y-axis values. If > 1, the
 magnitude of the scale amount will be added to the unit label 
 (e.g. scale_amount = 1000 with a unit label of "1000s of lbs" will result
  in "thousands of 1000s of lbs")

Default: 1}

\item{model_name}{String. Name of the model that will be present in the legend.

Options: Including, but not limited to: "msy", "unfished", "target"}
}
\value{
A ggplot2 geom_hline object for a reference point that can be added
to a plot
}
\description{
Preformatted reference line
}
\examples{
\dontrun{
reference_line(dat, "biomass", "msy")
}
}
