% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockplotr-package.R
\docType{package}
\name{stockplotr-package}
\alias{stockplotr}
\alias{stockplotr-package}
\title{stockplotr: Tables and Figures for Stock Assessments}
\description{
Creates exploratory and finished tables and figures for stock assessment documents from U.S. stock assessment model outputs. This packages addresses parts of the stock assessment workflow that interprets outputs of stock assessment models as well as allows the analyst to create report ready tables and figures, reducing the need to create their own and format then when adding into a report. This package is intended to be used in conjuction with `asar`, a partially automated template for writing various stock assessment reports. Throughout development, we will be creating a set of standardized figures and tables for a stock assessment report, developing functions to produce a variety of diagonostic plots, and other helpful materials. The advantage of using this package over others is that it applies to a range of stock assessment model outputs and standardizes them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nmfs-ost/stockplotr}
  \item Report bugs at \url{https://github.com/nmfs-ost/stockplotr/issues}
}

}
\author{
\strong{Maintainer}: Samantha Schiano \email{samantha.schiano@noaa.gov} (\href{https://orcid.org/0009-0003-3744-6428}{ORCID})

Authors:
\itemize{
  \item Sophie Breitbart \email{sophie.breitbart@noaa.gov} (\href{https://orcid.org/0000-0001-9641-9786}{ORCID}) [contributor]
  \item Steve Saul \email{steven.saul@noaa.gov}
}

Other contributors:
\itemize{
  \item Kelli Johnson \email{kelli.johnson@noaa.gov} (\href{https://orcid.org/0000-0002-5149-451X}{ORCID}) [contributor]
  \item Megumi Oshima \email{megumi.oshima@noaa.gov} (\href{https://orcid.org/0000-0002-8249-1442}{ORCID}) [contributor]
}

}
\keyword{internal}
