% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_indices.R
\name{table_indices}
\alias{table_indices}
\title{Create Indices of Abundance Table}
\usage{
table_indices(dat, end_year = NULL, make_rda = FALSE, tables_dir = getwd())
}
\arguments{
\item{dat}{A data frame returned from \link[asar]{convert_output}}

\item{end_year}{last year of assessment}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{tables_dir}{The location of the folder containing the generated table
rda files ("tables") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Create table of observed annual indices of abundance plus error
stratified by fleet.
}
\description{
Create Indices of Abundance Table
}
\examples{
\dontrun{
table_indices(dat)

table_indices(
  dat,
  end_year = 2024,
  make_rda = TRUE,
  tables_dir = getwd()
)
}
}
