% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_weight_at_age.R
\name{table_weight_at_age}
\alias{table_weight_at_age}
\title{Weight at age table}
\usage{
table_weight_at_age(
  dat,
  unit_label = "mt",
  era = NULL,
  interactive = TRUE,
  group = NULL,
  method = "sum",
  digits = 4,
  scale_amount = 1,
  make_rda = FALSE,
  tables_dir = getwd()
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{String. Abbreviated weight units

Default: "mt"}

\item{era}{String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{group}{String. Single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{method}{String. Method for summarizing data when group
is set to "none".

Default: "sum"

Options: "sum" or "mean"}

\item{digits}{Number. Numeric value indicating the number of digits values in the
table will be rounded to.

Default: 2}

\item{scale_amount}{Number. A number to scale the y-axis values.

Default: 1}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{tables_dir}{Path. The location of the folder containing the generated table
rda files ("tables") that will be created if the argument `make_rda` = TRUE.

Default: the working directory (`getwd()`)}
}
\value{
A report-ready table of weight by age class and year.
}
\description{
Weight at age table
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a [gt::gt()] object or export an rda object
containing a gt-based table, caption, and LaTeX-based table.
}
\examples{
table_weight_at_age(stockplotr::example_data)

table_weight_at_age(
  dat = stockplotr::example_data,
  unit_label = "grams"
)
}
\seealso{
[convert_output()], [filter_data()], [process_table()], [export_kqs()], [insert_kqs()], [create_rda()]
}
