% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_table.R
\name{theme_table}
\alias{theme_table}
\title{Add NOAA formatting to a table}
\usage{
theme_table(x)
}
\arguments{
\item{x}{Object. Table from \pkg{gt}, \pkg{flextable}, or \pkg{kableExtra}}
}
\value{
Add the standard formatting for stock assessment reports for any table.
}
\description{
Add NOAA formatting to a table
}
\details{
Currently, the function can format table objects from:
\href{https://ardata-fr.github.io/flextable-book/}{flextable} (\pkg{flextable} package, when installed),
\href{https://cran.r-project.org/web/packages/gt/index.html}{gt} (\pkg{gt}), and
\href{https://cran.r-project.org/web/packages/kableExtra/index.html}{kableExtra} (\pkg{kableExtra}).
}
\examples{
theme_table(cars |> gt::gt())
theme_table(cars |> kableExtra::kable())
}
