% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rda.R
\name{write_captions}
\alias{write_captions}
\title{Write captions and alternative text}
\usage{
write_captions(dat, dir = getwd(), year = format(Sys.Date(), "\%Y"))
}
\arguments{
\item{dat}{A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}. 

If inputting a list of tibbles, the first tibble's reference point defined 
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{dir}{Directory where the output captions and alt text file should be
saved. Defaults to working directory.}

\item{year}{the last year of the data or the current year this function is
being performed. Defaults to the current year.}
}
\value{
Exports .csv with captions and alt text for figures and tables
that contain key quantities (e.g., an assessment's start year) that
are automatically extracted from the converted model results file.
}
\description{
Function to create captions and alternative text that contain
key quantities from the model results file.
}
\examples{
\dontrun{
write_captions(dat,
  dir = getwd(),
  year = 2025
)
}
}
