% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_table.R
\name{check_label_differences}
\alias{check_label_differences}
\title{Create loop to test for differences in column values}
\usage{
check_label_differences(dat, index_variables, id_group = NULL)
}
\arguments{
\item{dat}{Data frame. Input data into process_table}

\item{index_variables}{Character vector. The index_variables vector created within process_table}

\item{id_group}{String. The identifying index variable as a string

Default: NULL}
}
\description{
Create loop to test for differences in column values
}
