% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_output.R
\name{convert_output}
\alias{convert_output}
\title{Convert Output}
\usage{
convert_output(file, model = NULL, fleet_names = NULL, save_dir = NULL)
}
\arguments{
\item{file}{Path. Assessment model output file path}

\item{model}{String. Assessment model used in evaluation

Default: NULL

Options: "ss3", "bam", "fims", "rceattle"}

\item{fleet_names}{Character vector. Names of fleets in the assessment model as
 shortened in the output file. If fleet names are not properly read, then
 indicate the fleets names as an acronym in a vector

 Default: NULL}

\item{save_dir}{Path. File path to save the converted output file.

Default: NULL}
}
\value{
A reformatted and standardized version of assessment model results
        for application in building a stock assessment reports and to easily
        adapt results among regional assessments.
}
\description{
Format stock assessment output files to a standardized format.
}
\details{
The resulting object is simply a transformed and machine readable
version of a model output file. Converted data frame is always returned.
It will also be saved if save_dir is not NULL.
}
\examples{
\dontrun{
convert_output(
  file = here::here("model1", "Report.sso"),
  model = "ss3",
  fleet_names = c("TWL", "NONTWL"),
  save_dir = here::here("standard_output.rda")
)
}
}
\author{
Samantha Schiano
}
