% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rda.R
\name{create_rda}
\alias{create_rda}
\title{Create the rda package for a plot or table}
\usage{
create_rda(
  object,
  topic_label,
  fig_or_table,
  dat,
  dir = getwd(),
  ref_line = "msy",
  scale_amount = 1,
  table_df = NULL
)
}
\arguments{
\item{object}{Object. Table or plot object}

\item{topic_label}{String. Name of the object}

\item{fig_or_table}{String. Indicates whether the object is a "table" or "figure"}

\item{dat}{Data frame. Data which will fill in captions and
alternative text for the object}

\item{dir}{Path. Directory to where the rda will be saved

Default: the working directory (`getwd()`)}

\item{ref_line}{String. Reference line value

Default: "msy"

Options: Including, but not limited to: "msy", "target", "unfished"}

\item{scale_amount}{Number. A number describing how much to scale down the quantities
shown on the y axis. For example, scale_amount = 100 would scale down a value
from 500,000 --> 5,000. This scale will be reflected in the y axis label.

Default: 1}

\item{table_df}{Data frame. The data frame that the table will be made into for purposes
of exporting a latex formatted table.}

\item{year}{Number. Assessment year

Default: the current year}
}
\value{
An rda package for a plot or table object. Requires an
object from the R environment such as a ggplot or flextable object.
}
\description{
Create the rda package for a plot or table
}
\examples{
\dontrun{
create_rda(
  object = my_plot,
  topic_label = "my_plot",
  fig_or_table = "figure",
  dat = my_data,
  dir = "path/to/save"
)
}
}
\seealso{
[export_rda()], [extract_caps_alttext()]
}
