% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rda.R
\name{create_rda}
\alias{create_rda}
\title{Create the rda package for a plot or table}
\usage{
create_rda(
  object,
  topic_label,
  fig_or_table,
  dat,
  dir = getwd(),
  year = format(as.POSIXct(Sys.Date(), format = "\%YYYY-\%mm-\%dd"), "\%Y"),
  ref_line = "msy",
  ref_point = "msy",
  scale_amount = 1,
  unit_label = "mt",
  table_df = NULL
)
}
\arguments{
\item{object}{Table or plot object}

\item{topic_label}{A string that names the object}

\item{fig_or_table}{A string identifying if the object is a "table" or "figure"}

\item{dat}{Data frame containing data which will fill in captions and
alternative text for the object}

\item{dir}{Directory to where the rda will be saved}

\item{year}{Default to current year}

\item{ref_line}{Reference line value such as "msy", "target" or "unfished"}

\item{ref_point}{Reference point value such as "msy", "target" or "unfished"}

\item{scale_amount}{A number describing how much to scale down the quantities
shown on the y axis. For example, scale_amount = 100 would scale down a value
from 500,000 --> 5,000. This scale will be reflected in the y axis label.}

\item{unit_label}{A string containing a unit label for the y-axis}

\item{table_df}{The data frame that the table will be made into for purposes 
of exporting a latex formatted table.}
}
\value{
Create an rda package for a plot or table object. Requires an
object from the R environment such as a ggplot or flextable object.
}
\description{
Create the rda package for a plot or table
}
\examples{
\dontrun{
create_rda(
  object = my_plot,
  topic_label = "my_plot",
  fig_or_table = "figure",
  dat = my_data,
  dir = "path/to/save"
)
}
}
