% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rda.R
\name{export_rda}
\alias{export_rda}
\title{Export a figure or table to rda}
\usage{
export_rda(
  object = NULL,
  caps_alttext = NULL,
  figures_tables_dir = NULL,
  topic_label = NULL,
  fig_or_table = NULL,
  latex_table = NULL
)
}
\arguments{
\item{object}{object. The final figure (ggplot) or table (flextable) object.

Default: NULL}

\item{caps_alttext}{data frame or list. The object containing a figure's caption and alternative
text, in a list, or a table's caption, likely generated with
stockplotr::extract_caps_alttext().

Default: NULL}

\item{figures_tables_dir}{path. If the user has already created folders containing
figures and tables ("figures" and "tables"), figures_tables_dir represents
the location of these folders. Otherwise, these two folders will be created
automatically, then used to store the exported rda files.

Default: NULL}

\item{topic_label}{string. A string that describes a figure or table's label. These
labels are found in the "label" column of the "captions_alt_text.csv" file
and are used to link the figure or table with its caption/alt text.

Default: NULL}

\item{fig_or_table}{string. A string describing whether the plot is a figure or table.

Default: NULL

Options: "figure", "table"}

\item{latex_table}{table object. The object containing a LaTeX-based table.

Default: NULL}
}
\value{
An rda file with a figure's ggplot, caption, and alternative text, or
a table's gt-based table, caption, and LaTeX-based table.
}
\description{
Export a figure/table, and its caption and alternative text, to an rda object.
Typically used after stockplotr::extract_caps_alttext().
}
\examples{
\dontrun{
export_rda(
  final = final_table_object,
  caps_alttext = caps_alttext_object,
  figures_tables_dir = here::here(),
  topic_label = "bnc",
  fig_or_table = "table",
  latex_table = "latex_table"
)

export_rda(
  final = final_figure_object,
  caps_alttext = another_caps_alttext_object,
  figures_tables_dir = "my_figures_tables_dir",
  topic_label = "landings",
  fig_or_table = "figure"
)
}
}
\seealso{
[create_rda()], [extract_caps_alttext()]
}
