% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_key_quantity.R
\name{extract_key_quantity}
\alias{extract_key_quantity}
\title{Extract key quantities}
\usage{
extract_key_quantity(dat, kq)
}
\arguments{
\item{dat}{A data frame or names list of data frames returned from
\link[asar]{convert_output}. The first data frame in the list is used in
calculation of a reference line if one is present}

\item{kq}{Key quantity that will be extracted. Possibilities include: _____________________________________________}
}
\value{
A string containing a value for a chosen key quantity.
}
\description{
Extract key quantities
}
\examples{
extract_key_quantity(
  dat = stockplotr:::example_data,
  kq = "B.start.year"
)
}
