% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sis_data.R
\name{extract_sis_data}
\alias{extract_sis_data}
\title{Extract data from model results to send to SIS}
\usage{
extract_sis_data(
  sis_data_dir = getwd(),
  key_quantities_dir = getwd(),
  figures_tables_dir = getwd()
)
}
\arguments{
\item{sis_data_dir}{Path. Location to save the sis_assmt_template.csv and
sis_ts_template.csv files or, if present, the location of the
existing sis_assmt_template.csv file.

Default: The working directory.}

\item{key_quantities_dir}{Path. Location of the existing key_quantities.csv file.

Default: The working directory.}

\item{figures_tables_dir}{Path. Location of the existing 'figures' and 'tables' directories.

Default: The working directory.}
}
\description{
Semi-automate the extraction of key quantities from model results for eventual transmittance to SIS via `asar::export_to_sis()`.
}
\details{
This function acts within the following workflow:

1. When a stock assessment is scheduled to conclude, SIS will automatically
assign an attachment or prompt containing metadata and identifiers.
2. The user will open two csv files containing placeholders for all of the data required by SIS: sis_assmt_template.csv (assessment summary data) and sis_ts_template.csv (time series data). There are three ways to obtain these files:
2a. Run `stockplotr::extract_sis_data()`, which will generate, populate, and export the templates with data originating from a converted model results file.
At least one of the following rdas must be present in the "figures" and/or
"tables" folders: biomass_figure.rda, spawning_biomass_figure.rda,
abundance_at_age_figure.rda, fishing_mortality_figure.rda, index_figure.rda,
total_catch_table.rda.
2b. Generate blank files by running `asar::create_blank_sis()`.
2c. Locate blank files in the "report" folder generated by running `asar::create_template()`.
3. The user will add the remaining necessary data into the csv files, ensuring that all required fields are completed.
4. Run `export_to_sis()`, which will format and upload this data to a specific Google Drive folder.
5. The uploaded contents will be resubmitted to SIS to finalize the record.

Data are extracted from the `.rda` files created by setting `make_rda` = TRUE
in the plotting functions:

\strong{Biomass}: `stockplotr::plot_biomass()`

\strong{Spawning Biomass}: `stockplotr::plot_spawning_biomass()`

\strong{Recruitment at Age 1}: `stockplotr::plot_abundance_at_age()`

\strong{Fishing Mortality}: `stockplotr::plot_fishing_mortality()`

\strong{Index of Abundance}: `stockplotr::plot_index()`

\strong{Catch}: `stockplotr::plot_total_catch()`
}
\examples{
\dontrun{
extract_sis_data(
  sis_data_dir = getwd(),
  key_quantities_dir = "my_dir",
  figures_tables_dir = "my_other_dir"
)
}

}
