% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sis_data.R
\name{extract_sis_data}
\alias{extract_sis_data}
\title{Extract data from model results to send to SIS}
\usage{
extract_sis_data(
  sis_data_dir = getwd(),
  key_quantities_dir = getwd(),
  figures_tables_dir = getwd()
)
}
\arguments{
\item{sis_data_dir}{Path. Location of the existing sis_assmt_template.csv 
file, or, if absent, where a new version of sis_assmt_template.csv and
sis_ts_template.csv should be saved.

Default: The working directory.}

\item{figures_tables_dir}{Path. Location of the existing 'figures' and 'tables' directories.

Default: The working directory.}

\item{kq_dir}{Path. Location of the existing key_quantities.csv file, or, if absent, where a new version of the file should be saved.

Default: The working directory.}
}
\description{
Semi-automate the extraction of key quantities from model results for eventual transmittance to SIS via `asar::export_to_sis()`.
}
\details{
This function acts within the following workflow:

1. When a stock assessment is scheduled to conclude, SIS will generate an
 attachment or prompt containing metadata and identifiers.
2. The user will open two csv files containing placeholders for all of the data required by SIS: sis_assmt_template.csv (assessment summary data) and sis_ts_template.csv (time series data). There are three ways to obtain these files:
2a. Generate the files by running `asar::create_blank_sis()`
2b. Locate the files in the "report" folder generated by running `asar::create_template()`.
2c. Run `stockplotr::extract_sis_data()`, which will populate the templates with data originating from a converted model results file.
3. The user will add the remaining necessary data into the csv files, ensuring that all required fields are completed.
4. Run `export_to_sis()`, which will format and upload this data to a specific Google Drive folder.
5. The uploaded contents will be resubmitted to SIS to finalize the record.
}
\examples{
\dontrun{
extract_sis_data(
  sis_data_dir = getwd()
)
}

}
