% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{filter_data}
\alias{filter_data}
\title{Filter data for input into aesthetics for ggplot2}
\usage{
filter_data(
  dat,
  label_name,
  module = NULL,
  era = NULL,
  geom,
  group = NULL,
  facet = NULL,
  scale_amount = 1,
  interactive = TRUE
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{label_name}{String. Name of the label used to filter
the data.}

\item{module}{Character vector. (Optional) Module name found in `dat$module_name`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{era}{String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{geom}{String. Geom used for the plot.

Default: "line".

Options: "line", "point", or "area"}

\item{group}{String. Single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{Character vector. Column name or names used for faceting.

Default: NULL}

\item{scale_amount}{Number. A number to scale the y-axis values.

Default: 1}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}
}
\value{
a data frame that is preformatted for plotting with ggplot2.
}
\description{
Filter data for input into aesthetics for ggplot2
}
\examples{
\dontrun{
filter_data(dat, "biomass", "line", group = "fleet")
}
}
