% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{filter_data}
\alias{filter_data}
\title{Prep data for input into aesthetics for ggplot2}
\usage{
filter_data(
  dat,
  label_name,
  module = NULL,
  era = "time",
  geom,
  group = NULL,
  facet = NULL,
  scale_amount = 1,
  interactive = TRUE
)
}
\arguments{
\item{dat}{a data frame or list of data frames that contains the data to be
plotted.}

\item{label_name}{a string of the name of the label that is used to filter
the data.}

\item{module}{(Optional) A string indicating the linked module_name associated
with the label for the plot if known. Default is NULL. By default, the function
will select the most relevant module if more than 1 exists.}

\item{era}{A string naming the era of data such as historical ("early"), current ("time"), or 
projected ("fore") data if filtering should occur. Default is set to "time" which is 
the current time. To plot all data, set era to NULL.}

\item{geom}{Type of plot user wants to create. Options are "line", "point",
and "area".}

\item{group}{Selected grouping for the data. If you want to just summarize
the data across all factors, set group = "none".}

\item{facet}{a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.)}

\item{scale_amount}{A number describing how much to scale down the quantities
shown on the y axis.}

\item{interactive}{logical. If TRUE, the user will be prompted to select
a module_name when there was more than one found in the filtered dataset.}
}
\value{
a data frame that is preformatted for plotting with ggplot2.
}
\description{
Prep data for input into aesthetics for ggplot2
}
\examples{
\dontrun{
filter_data(dat, "biomass", "line", group = "fleet")
}
}
