% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_all_figs_tables.R
\name{html_all_figs_tables}
\alias{html_all_figs_tables}
\title{Make an html file with all figures and tables}
\usage{
html_all_figs_tables(rda_dir = getwd(), outdir = getwd())
}
\arguments{
\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}

\item{outdir}{The location of the folder containing the generated html and qmd
files ("all_tables_figures") that will be created. Default is the working directory.}
}
\value{
An html file with all existing figures and tables.
}
\description{
Show all figures and tables in an html file.
}
\examples{
\dontrun{
html_all_figs_tables()
html_all_figs_tables(rda_dir = getwd(),
                     outdir = getwd())
}
}
