% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_all_figs_tables.R
\name{html_all_figs_tables}
\alias{html_all_figs_tables}
\title{Make an html file with all figures and tables}
\usage{
html_all_figs_tables(figures_tables_dir = getwd())
}
\arguments{
\item{figures_tables_dir}{The location of the folder containing
figures and tables ("figures" and "tables").}
}
\value{
A folder ("all_tables_figures") in your working directory containing
html and qmd files that show all tables and figures.
}
\description{
Show all tables and figures in a single html file.
}
\examples{
\dontrun{
html_all_figs_tables()
html_all_figs_tables(figures_tables_dir = "my_figures_tables_dir")
}
}
