% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_aa}
\alias{plot_aa}
\title{Create "at-age" plot}
\usage{
plot_aa(
  dat,
  x = "year",
  y = "age",
  z = "estimate",
  label = "Abundance",
  xlab = "Year",
  ylab = "Age",
  facet = NULL,
  proportional = TRUE,
  ...
)
}
\arguments{
\item{dat}{filtered data frame from standard output file(s) preformatted for
the target label from \link[stockplotr]{filter_data}}

\item{x}{a string of the column name of data used to plot on the x-axis

Default: "year"}

\item{y}{a string of the column name of data used to plot on the y-axis

Default: "age"}

\item{z}{a string of the column name of data used to control the size of the
bubbles

Default: "estimate"}

\item{label}{a string of the label for the size of the bubbles

Default: "Abundance"}

\item{xlab}{a string of the x-axis label

Default: "Year"}

\item{ylab}{a string of the y-axis label

Default: "Age"}

\item{facet}{a string or vector of strings of a column that facets the data.
It is not recommended to include more than one facet due to the complexity 
of the plot.

Default: NULL

Options: Including, but not limited to: "sex", "area", "fleet"}

\item{proportional}{Set size of points relative to z when TRUE, point
size are relative to one another while when set to FALSE, point size
is relative to z

Default: `TRUE`}

\item{...}{inherited arguments from internal functions from
\link[ggplot2]{geom_point}}
}
\value{
Create a plot of abundance at age for a stock assessment report.
}
\description{
Create "at-age" plot
}
\examples{
\dontrun{
plot_aa(dat)
}
}
