% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_aa}
\alias{plot_aa}
\title{Create "at-age" plot}
\usage{
plot_aa(
  dat,
  x = "year",
  y = "age",
  z = "estimate",
  label = "Abundance",
  xlab = "Year",
  ylab = "Age",
  facet = NULL,
  proportional = TRUE,
  ...
)
}
\arguments{
\item{dat}{filtered data frame from standard output file(s) preformatted for
the target label from \link[stockplotr]{prepare_data}}

\item{x}{a string of the column name of data used to plot on the x-axis
(default is "year")}

\item{y}{a string of the column name of data used to plot on the y-axis
(default is "age")}

\item{z}{a string of the column name of data used to control the size of the
bubbles (default is "estimate")}

\item{label}{a string of the label for the size of the bubbles
(default is "Abundance")}

\item{xlab}{a string of the x-axis label (default is "Year")}

\item{ylab}{a string of the y-axis label (default is "Age")}

\item{facet}{a string or vector of strings of a column that facets the data
(e.g. "sex", "area", etc.). It is not recommended to include more than one
facet due to the complexity of the plot.}

\item{proportional}{Set size of points relative to z when TRUE, point
size are relative to one another while when set to FALSE, point size
is relative to z}

\item{...}{inherited arguments from internal functions from
\link[ggplot2]{geom_point}}
}
\value{
Create a plot of abundance at age for a stock assessment report.
}
\description{
Create "at-age" plot
}
\examples{
\dontrun{
plot_aa(dat)
}
}
