% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance_at_age.R
\name{plot_abundance_at_age}
\alias{plot_abundance_at_age}
\title{Plot Abundance (or Numbers) at Age (AAA or NAA)}
\usage{
plot_abundance_at_age(
  dat,
  unit_label = "fish",
  scale_amount = 1000,
  end_year = NULL,
  make_rda = FALSE,
  figures_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from \link[asar]{convert_output}}

\item{unit_label}{units for recruitment}

\item{scale_amount}{A number describing how much to scale down the abundance at
age. Please choose a value ranging from 1-1,000,000,000 (one billion) in orders
of magnitude (e.g., 1, 10, 100, 1000, etc.). For example, scale_amount = 100
would scale down a value from 500,000 --> 5,000 and would report abundance in
hundreds of fish (if "fish" was the unit_label). This scale will be reflected
in the legend label. The default is 1,000.}

\item{end_year}{last year of assessment}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{figures_dir}{The location of the folder containing the generated figure
rda files ("figures") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Plot total abundance (or numbers) at age from a stock assessment model as found in a NOAA
stock assessment report. Units of abundance can either be manually added
or will be extracted from the provided file if possible.
}
\description{
Plot Abundance (or Numbers) at Age (AAA or NAA)
}
\examples{
\dontrun{
plot_abundance_at_age(dat)

plot_abundance_at_age(
  dat,
  unit_label = "my_unit",
  scale_amount = 10,
  end_year = 2024,
  make_rda = TRUE,
  figures_dir = getwd()
)
}
}
