% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot Total Biomass}
\usage{
plot_biomass(
  dat,
  geom = "line",
  group = NULL,
  facet = NULL,
  ref_line = "msy",
  unit_label = "metric tons",
  module = NULL,
  scale_amount = 1,
  relative = FALSE,
  make_rda = FALSE,
  figures_dir = getwd(),
  interactive = TRUE,
  ...
)
}
\arguments{
\item{dat}{A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}. 
If inputting a list of tibbles, the first tibble's reference point defined 
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{geom}{A string stating the geom used for the plot. 
Default: "line".
Options: "line", "point", or "area"}

\item{group}{A string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.). 
Set group = "none" to summarize data over all indexing values.
Default: NULL}

\item{facet}{A string or vector of strings of a column name.
Default: NULL}

\item{ref_line}{A string specifying the type of reference you want to
compare biomass to. The default is `"msy"`, which looks for
`"biomass_msy"` in the `"label"` column of `dat`. The actual
searching in `dat` is case agnostic and will work with either upper- or
lower-case letters but you must use one of the options specified in the
default list to ensure that the label on the figure looks correct
regardless of how it is specified in `dat`. Other possibilities may include
"target", "MSY", and "unfished".}

\item{unit_label}{units for biomass}

\item{module}{(Optional) A string indicating the module_name found in `dat`.
Default: NULL
If the interactive and >1 module_name is found, user will select the 
module_name in the console. @seealso [filter_data()]}

\item{scale_amount}{A number to scale the y-axis values.
Default: 1}

\item{relative}{A logical value specifying to set y-axis values relative to 
the ref_line value.
Default: `FALSE`}

\item{make_rda}{A logical value indicating whether to save the object and 
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default: `FALSE`.}

\item{figures_dir}{A string indicating a path to the "figures" folder.
Default: `getwd()`
The folder is created within the path if it does not exist.}

\item{interactive}{A logical value indicating if the environment is interactive.
Default: `FALSE`}

\item{...}{Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
Plot total biomass from a stock assessment model as found in a NOAA
stock assessment report. Units of total biomass can either be manually added
or will be extracted from the provided file if possible. There are options #' to return a [ggplot2::ggplot()] object or export an rda object containing
associated caption and alternative text for the figure.
}
\description{
Plot Total Biomass
}
\examples{
plot_biomass(
  dat = stockplotr:::example_data,
  unit_label = "mt",
  ref_line = "target",
  scale_amount = 100,
  module = "TIME_SERIES",
  figures_dir = getwd()
)
plot_biomass(
  dat = stockplotr:::example_data,
  facet = "fleet",
  ref_line = "target",
  module = "TIME_SERIES",
  figures_dir = getwd()
)
}
