% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass_at_age.R
\name{plot_biomass_at_age}
\alias{plot_biomass_at_age}
\title{Plot Total Biomass at Age (BAA)}
\usage{
plot_biomass_at_age(
  dat,
  facet = NULL,
  unit_label = "mt",
  scale_amount = 1000,
  proportional = TRUE,
  make_rda = FALSE,
  figures_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from \link[asar]{convert_output}}

\item{facet}{a string or vector of strings of column(s) that 
groups the data (e.g. "fleet", "sex", "area", etc.).}

\item{unit_label}{units for abundance}

\item{scale_amount}{A number describing how much to scale down the biomass at
age. For example, scale_amount = 100 would scale down a value
from 500,000 --> 5,000. This scale will be reflected in the legend label.}

\item{proportional}{Set size of points relative to z when TRUE, point
size are relative to one another while when set to FALSE, point size
is relative to z}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the .rda will be exported to the folder indicated in the argument "rda_dir".
Default is FALSE.}
}
\value{
Plot total biomass at age from a stock assessment model as found in a NOAA
stock assessment report. Units of total biomass can either be manually added
or will be extracted from the provided file if possible.
}
\description{
Plot Total Biomass at Age (BAA)
}
\examples{
\dontrun{
plot_biomass_at_age(dat)

plot_biomass_at_age(
  dat,
  unit_label = "my_unit",
  scale_amount = 100,
  end_year = 2024,
  make_rda = TRUE,
  figures_dir = getwd()
)
}
}
