% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_comp.R
\name{plot_catch_comp}
\alias{plot_catch_comp}
\title{Plot catch composition by fleet}
\usage{
plot_catch_comp(
  dat,
  unit_label = "metric tons",
  end_year = NULL,
  make_rda = FALSE,
  rda_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{unit_label}{indicate the name of the units of catch as to label the axis}

\item{end_year}{last year of assessment}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the .rda will be exported to the folder indicated in the argument "rda_dir".
Default is FALSE.}

\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
A plot ready for a stock assessment report of catch composition by fleet.
This plot is made only when landings are explicitly named in the output file.
The current plot function does not combine all sources of catch.
}
\description{
Plot catch composition by fleet
}
