% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_error}
\alias{plot_error}
\title{Create plot with error}
\usage{
plot_error(
  dat,
  x = "year",
  y = "estimate",
  geom = "point",
  group = NULL,
  facet = NULL,
  xlab = "Year",
  ylab = NULL,
  ...
)
}
\arguments{
\item{dat}{filtered data frame from standard output file(s) pre-formatted for
the target label from \link[stockplotr]{prepare_data}}

\item{x}{a string of the column name of data used to plot on the x-axis (default 
is "year")}

\item{y}{a string of the column name of data used to plot on the y-axis (default 
is "estimate")}

\item{geom}{type of geom to use for plotting found in ggplot2 (e.g. "point", 
"line", etc.). Default is "line". Other options are "point" and "area".}

\item{group}{a string of a single column that groups the data (e.g. "fleet", 
"sex", "area", etc.). Currently can only have one level of grouping.}

\item{facet}{a string or vector of strings of a column that facets the data 
(e.g. "year", "area", etc.)}

\item{xlab}{a string of the x-axis label (default is "Year")}

\item{ylab}{a string of the y-axis label. If NULL, it will be set to the name
of `y`.}

\item{...}{inherited arguments from internal functions from ggplot2::geom_xx}
}
\description{
Create plot with error
}
