% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fishing_mortality.R
\name{plot_fishing_mortality}
\alias{plot_fishing_mortality}
\title{Plot fishing mortality timeseries}
\usage{
plot_fishing_mortality(
  dat,
  geom = "line",
  group = NULL,
  facet = NULL,
  ref_line = "msy",
  era = NULL,
  module = NULL,
  make_rda = FALSE,
  figures_dir = getwd(),
  interactive = TRUE,
  ...
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{geom}{String. A string stating the geom used for the plot.

Default: "line".

Options: "line", "point", or "area"}

\item{group}{String. A string of a single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{Character vector. A string or vector of strings of a column name.

Default: NULL}

\item{ref_line}{String. A string specifying a reference point name.

Default: "target"

Options: (including, but not limited to) "target", "msy", and "unfished"
If the reference point is not found in the data, set ref_line = c("name" = value).}

\item{era}{String. A string naming the era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{module}{Character vector. (Optional) A string indicating the module_name found in `dat`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{Path. A string indicating a path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{...}{Dots. Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
A plot showing fishing mortality over time.
}
\description{
Plot fishing mortality timeseries
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a `ggplot2` object or export an .rda object
containing associated caption and alternative text for the figure.
}
\examples{
plot_fishing_mortality(
  dat = stockplotr:::example_data,
  ref_line = c("target" = 0.2),
  group = "fleet",
  interactive = FALSE,
  module = "TIME_SERIES"
)
plot_fishing_mortality(
  dat = stockplotr:::example_data,
  ref_line = c("target" = 0.2),
  group = "none",
  interactive = FALSE,
  module = "TIME_SERIES"
)
}
\seealso{
[convert_output()], [filter_data()], [process_data()], [plot_timeseries()], [reference_line()], [export_kqs()], [insert_kqs()], [create_rda()]
}
