% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indices.R
\name{plot_indices}
\alias{plot_indices}
\title{Plot Index of Abundance}
\usage{
plot_indices(
  dat,
  unit_label = "",
  group = NULL,
  facet = "fleet",
  interactive = TRUE,
  module = NULL,
  focus = NULL,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{units for index of abundance/CPUE}

\item{group}{A string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.).

Set group = "none" to summarize data over all indexing values.

Default: NULL}

\item{facet}{a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.) "fleet" is always added on to any faceting selections}

\item{interactive}{A logical value indicating if the environment is interactive.

Default: `FALSE`}

\item{module}{(Optional) A string indicating the module_name found in `dat`.

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{focus}{a string or vector of strings indicating how data should be
filtered. (i.e. select names of fleets to zoom into the plot)}

\item{make_rda}{A logical value indicating whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{A string indicating a path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}

\item{...}{Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
Plot the expected and predicted indices as indicated from a standard
assessment model output file.
}
\description{
Plot Index of Abundance
}
\examples{
plot_indices(
  dat = stockplotr:::example_data,
  unit_label = "fish/hr",
  interactive = FALSE
)

}
