% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indices.R
\name{plot_indices}
\alias{plot_indices}
\title{Plot Index of Abundance}
\usage{
plot_indices(
  dat,
  unit_label = "",
  group = NULL,
  facet = "fleet",
  interactive = TRUE,
  module = NULL,
  focus = NULL,
  make_rda = FALSE,
  figures_dir = NULL,
  ...
)
}
\arguments{
\item{dat}{A data frame or names list of data frames returned from
\link[asar]{convert_output}. The first data frame in the list is used in
calculation of a reference line if one is present}

\item{unit_label}{units for index of abundance/CPUE}

\item{group}{a string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.). Currently can only have one level of grouping.}

\item{facet}{a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.) "fleet" is always added on to any faceting selections}

\item{interactive}{TRUE/FALSE; indicate whether the environment in which the
function is operating  is interactive. This bypasses some options for
filtering when preparing data for the plot. Default is FALSE.}

\item{module}{(Optional) A string indicating the linked module_name associated
with the label for the plot if known. Default is NULL. By default, the function
will select the most relevant module if more than 1 exists.}

\item{focus}{a string or vector of strings indicating how data should be 
filtered. (i.e. select names of fleets to zoom into the plot)}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{figures_dir}{The location of the folder containing the generated figure
rda files ("figures") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}

\item{...}{Arguments called from ggplot2::geom_line or ggplot2::geom_point}
}
\value{
Plot the expected and predicted indices as indicated from a standard 
assessment model output file.
}
\description{
Plot Index of Abundance
}
\examples{
\dontrun{
plot_indices(
  petrale,
  unit_label = "fish/hr",
  interactive= TRUE,
  module = NULL,
  make_rda = FALSE,
  figures_dir = getwd()
)
}
}
