% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landings.R
\name{plot_landings}
\alias{plot_landings}
\title{Plot observed landings}
\usage{
plot_landings(
  dat,
  unit_label = "mt",
  geom = "line",
  group = NULL,
  facet = NULL,
  lbs = FALSE,
  era = NULL,
  scale_amount = 1,
  module = NULL,
  interactive = TRUE,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{string. A string specifying spawning biomass unit.

Default: "mt"}

\item{geom}{string. A string stating the geom used for the plot.

Default: "line".

Options: "line", "point", or "area"}

\item{group}{string. A string of a single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{character vector. A string or vector of strings of a column name.

Default: NULL}

\item{lbs}{logical. A logical value indicating whether to convert the y-axis values from
kilograms to pounds. The default units match the default in the
unit_label argument - 'mt'.

Default: `FALSE`}

\item{era}{string. A string naming the era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{scale_amount}{number. A number to scale the y-axis values.

Default: 1}

\item{module}{character vector. (Optional) A string indicating the module_name found in `dat`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{interactive}{logical. A logical value indicating if the environment is interactive.

Default: `FALSE`}

\item{make_rda}{logical. A logical value indicating whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{path. A string indicating a path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}

\item{...}{dots. Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
A plot showing cumulative landings over time.
}
\description{
Plot observed landings
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a `ggplot2` object or export an .rda object
containing associated caption and alternative text for the figure.
}
\examples{
plot_landings(
  dat = stockplotr:::example_data,
  unit_label = "metric tons",
  group = "fleet",
  interactive = FALSE,
  make_rda = FALSE
)
plot_landings(
  dat = stockplotr:::example_data,
  unit_label = "metric tons",
  facet = "fleet",
  interactive = FALSE,
  make_rda = FALSE
)
}
\seealso{
[convert_output()], [filter_data()], [process_data()], [plot_timeseries()], [export_kqs()], [insert_kqs()], [create_rda()]
}
