% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_natural_mortality.R
\name{plot_natural_mortality}
\alias{plot_natural_mortality}
\title{Plot natural mortality (M) at age}
\usage{
plot_natural_mortality(
  dat,
  group = NULL,
  facet = NULL,
  era = NULL,
  interactive = TRUE,
  module = NULL,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}. 

If inputting a list of tibbles, the first tibble's reference point defined 
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{group}{A string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.). 

Set group = "none" to summarize data over all indexing values.

Default: NULL}

\item{facet}{A string or vector of strings of a column name.

Default: NULL}

\item{era}{A string naming the era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{interactive}{A logical value indicating if the environment is interactive.

Default: `FALSE`}

\item{module}{(Optional) A string indicating the module_name found in `dat`.

Default: NULL

If the interactive and >1 module_name is found, user will select the 
module_name in the console. @seealso [filter_data()]}

\item{make_rda}{A logical value indicating whether to save the object and 
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{A string indicating a path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}

\item{...}{Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
Plot natural mortality at age from a stock assessment model as
found in a NOAA stock assessment report.
}
\description{
Plot natural mortality (M) at age
}
\examples{
plot_natural_mortality(
  dat = stockplotr:::example_data,
  module = "Natural_Mortality",
  interactive = FALSE
)
}
