% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_natural_mortality.R
\name{plot_natural_mortality}
\alias{plot_natural_mortality}
\title{Plot natural mortality (M) at age}
\usage{
plot_natural_mortality(
  dat,
  group = NULL,
  facet = NULL,
  era = NULL,
  interactive = TRUE,
  module = NULL,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{A data frame or names list of data frames (input as `list()`)
returned from \link[asar]{convert_output}. The first data frame in the list 
is used in calculation of a reference line if one is present}

\item{group}{a string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.). Currently can only have one level of grouping. If you want to just summarize
the data across all factors, set group = "none". In the case there is only
one unique value of the grouping and/or NA is available, the function will
default to the NAs and in some cases the grouping when there is not an equal
number of rows for NA and the grouped data.}

\item{facet}{a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.)}

\item{era}{a string naming the era of data such as historical ("early"), current ("time"), or 
projected ("fore") data if filtering should occur. Default is set to "time" which is 
the current time. To plot all data, set era to NULL.}

\item{interactive}{TRUE/FALSE; indicate whether the environment in which the
function is operating  is interactive. This bypasses some options for
filtering when preparing data for the plot. Default is FALSE.}

\item{module}{(Optional) A string indicating the linked module_name associated
with the label for the plot if known. Default is NULL. By default, the function
will select the most relevant module if more than 1 exists.}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{figures_dir}{The location of the folder containing the generated figure
rda files ("figures") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}

\item{...}{Arguments called from ggplot2::geom_line or ggplot2::geom_point}
}
\value{
Plot natural mortality at age from a stock assessment model as 
found in a NOAA stock assessment report.
}
\description{
Plot natural mortality (M) at age
}
\examples{
plot_natural_mortality(
  dat = stockplotr:::example_data,
  module = "Natural_Mortality",
  interactive = FALSE
)
}
