% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_obsvpred}
\alias{plot_obsvpred}
\title{Plot observed vs. predicted data}
\usage{
plot_obsvpred(
  dat,
  x = "year",
  y = "estimate",
  observed_label = "observed",
  predicted_label = "predicted",
  geom = "line",
  xlab = "Year",
  ylab = NULL,
  group = NULL,
  facet = NULL,
  ...
)
}
\arguments{
\item{dat}{Data frame. Filtered data frame from standard output file(s) preformatted for
the target label from \link[stockplotr]{filter_data}}

\item{x}{String. Column name of data used to plot on the x-axis

Default: "year"}

\item{y}{String. Column name of data used to plot on the y-axis

Default: "estimate"}

\item{observed_label}{String. Label used to filter the observed data

Default: "observed"}

\item{predicted_label}{String. Label used to filter the predicted data
Default: "predicted"}

\item{geom}{String. Type of geom to use for plotting found in ggplot2 (e.g. "point",
"line", etc.).

Default: "line"

Options: "point" and "area"}

\item{xlab}{String. X-axis label

Default: "Year"}

\item{ylab}{String. Y-axis label. If NULL, it will be set to the name
 of `y`.

Default: NULL}

\item{group}{String. Single column that groups the data.
Currently can only have one level of grouping.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{Character vector. Column name or names used for faceting
(e.g. "year", "area", etc.)

Default: NULL}

\item{...}{Inherited arguments from internal functions from ggplot2::geom_xx}
}
\value{
A plot of observed vs. predicted data for a stock assessment report.
}
\description{
Plot observed vs. predicted data
}
