% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_obsvpred}
\alias{plot_obsvpred}
\title{Plot observed vs. predicted data}
\usage{
plot_obsvpred(
  dat,
  x = "year",
  y = "estimate",
  observed_label = "observed",
  predicted_label = "predicted",
  geom = "line",
  xlab = "Year",
  ylab = NULL,
  group = NULL,
  facet = NULL,
  ...
)
}
\arguments{
\item{dat}{data frame. filtered data frame from standard output file(s) preformatted for
the target label from \link[stockplotr]{filter_data}}

\item{x}{string. a string of the column name of data used to plot on the x-axis

Default: "year"}

\item{y}{string. a string of the column name of data used to plot on the y-axis

Default: "estimate"}

\item{observed_label}{string. a string of the label used to filter the observed data

Default: "observed"}

\item{predicted_label}{string. a string of the label used to filter the predicted data
Default: "predicted"}

\item{geom}{string. type of geom to use for plotting found in ggplot2 (e.g. "point",
"line", etc.).

Default: "line"

Options: "point" and "area"}

\item{xlab}{string. a string of the x-axis label

Default: "Year"}

\item{ylab}{string. a string of the y-axis label. If NULL, it will be set to the name
 of `y`.

Default: NULL}

\item{group}{string. a string of a single column that groups the data.
Currently can only have one level of grouping.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{character vector. a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.)

Default: NULL}

\item{...}{dots. inherited arguments from internal functions from ggplot2::geom_xx}
}
\value{
A plot of observed vs. predicted data for a stock assessment report.
}
\description{
Plot observed vs. predicted data
}
