% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_obsvpred}
\alias{plot_obsvpred}
\title{Plot observed vs. predicted data}
\usage{
plot_obsvpred(
  dat,
  x = "year",
  y = "estimate",
  observed_label = "observed",
  predicted_label = "predicted",
  geom = "line",
  xlab = "Year",
  ylab = NULL,
  group = NULL,
  facet = NULL,
  ...
)
}
\arguments{
\item{dat}{filtered data frame from standard output file(s) preformatted for
the target label from \link[stockplotr]{filter_data}}

\item{x}{a string of the column name of data used to plot on the x-axis (default
is "year")}

\item{y}{a string of the column name of data used to plot on the y-axis (default
is "estimate")}

\item{observed_label}{a string of the label used to filter the observed data. Default is "observed".}

\item{predicted_label}{a string of the label used to filter the predicted data. Default is "predicted".}

\item{geom}{type of geom to use for plotting found in ggplot2 (e.g. "point",
"line", etc.). Default is "line". Other options are "point" and "area".}

\item{xlab}{a string of the x-axis label (default is "Year")}

\item{ylab}{a string of the y-axis label. If NULL, it will be set to the name
of `y`.}

\item{group}{a string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.). Currently can only have one level of grouping.}

\item{facet}{a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.)}

\item{...}{inherited arguments from internal functions from ggplot2::geom_xx}
}
\value{
Create a plot of observed vs. predicted data for a stock assessment report.
}
\description{
Plot observed vs. predicted data
}
