% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot Recruitment}
\usage{
plot_recruitment(
  dat,
  unit_label = "mt",
  scale_amount = 1,
  era = NULL,
  group = NULL,
  facet = NULL,
  interactive = TRUE,
  module = NULL,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{String. Units for recruitment

Default: "mt"}

\item{scale_amount}{Number. A number to scale the y-axis values.

Default: 1}

\item{era}{String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{group}{String. Single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{facet}{Character vector. Column name or names used for faceting.

Default: NULL}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{module}{Character vector. (Optional) Module name found in `dat$module_name`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{Path. Path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}

\item{...}{Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
A plot showing recruitment over time.
}
\description{
Plot Recruitment
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a `ggplot2` object or export an .rda object
containing associated caption and alternative text for the figure.
}
\examples{
plot_recruitment(
  dat = stockplotr:::example_data,
  unit_label = "metric tons",
  scale_amount = 100,
  interactive = TRUE,
  module = "TIME_SERIES",
  make_rda = FALSE
)
plot_recruitment(
  dat = stockplotr:::example_data,
  era = "fore",
  module = "TIME_SERIES",
  make_rda = FALSE
)
}
\seealso{
[convert_output()], [filter_data()], [process_data()], [plot_timeseries()], [export_kqs()], [insert_kqs()], [create_rda()]
}
