% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot Recruitment}
\usage{
plot_recruitment(
  dat,
  unit_label = "mt",
  scale_amount = 1,
  era = "time",
  group = NULL,
  facet = NULL,
  interactive = TRUE,
  module = NULL,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{units for recruitment}

\item{scale_amount}{A number to scale the y-axis values.

Default: 1}

\item{era}{A string naming the era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{group}{A string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.).

Set group = "none" to summarize data over all indexing values.

Default: NULL}

\item{facet}{A string or vector of strings of a column name.

Default: NULL}

\item{interactive}{A logical value indicating if the environment is interactive.

Default: `FALSE`}

\item{module}{(Optional) A string indicating the module_name found in `dat`.

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{make_rda}{A logical value indicating whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{A string indicating a path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}
}
\value{
Plot recruitment over time from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}). There are options to return a
[ggplot2::ggplot()] object or export an rda object containing associated
caption and alternative text for the figure.
}
\description{
Plot Recruitment
}
\examples{
plot_recruitment(
  dat = stockplotr:::example_data,
  unit_label = "metric tons",
  scale_amount = 100,
  interactive = TRUE,
  module = "TIME_SERIES",
  make_rda = FALSE
)
plot_recruitment(
  dat = stockplotr:::example_data,
  era = "fore",
  module = "TIME_SERIES",
  make_rda = FALSE
)
}
