% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment_deviations.R
\name{plot_recruitment_deviations}
\alias{plot_recruitment_deviations}
\title{Plot Recruitment Deviations}
\usage{
plot_recruitment_deviations(
  dat,
  module = NULL,
  era = "time",
  interactive = TRUE,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{A data frame or names list of data frames (input as `list()`)
returned from \link[asar]{convert_output}. The first data frame in the list 
is used in calculation of a reference line if one is present}

\item{module}{(Optional) A string indicating the linked module_name associated
with the label for the plot if known. Default is NULL. By default, the function
will select the most relevant module if more than 1 exists.}

\item{era}{a string naming the era of data such as historical ("early"), current ("time"), or 
projected ("fore") data if filtering should occur. Default is set to "time" which is 
the current time. To plot all data, set era to NULL.}

\item{interactive}{TRUE/FALSE; indicate whether the environment in which the
function is operating  is interactive. This bypasses some options for
filtering when preparing data for the plot. Default is FALSE.}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{figures_dir}{The location of the folder containing the generated figure
rda files ("figures") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}

\item{...}{Arguments called from ggplot2::geom_line or ggplot2::geom_point}
}
\value{
Plot recruitment deviations relative to one over time from an
assessment model output file translated to a standardized output (\link[asar]{convert_output}). There are
options to return a `ggplot2` object or export an .rda object containing
associated caption and alternative text for the figure.
}
\description{
Plot Recruitment Deviations
}
\examples{
plot_recruitment_deviations(
  dat = stockplotr:::example_data,
  interactive = FALSE,
  size = 1.5
)
plot_recruitment_deviations(
  dat = stockplotr:::example_data,
  era = "early",
  interactive = FALSE,
  shape = 2
)
}
