% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment_deviations.R
\name{plot_recruitment_deviations}
\alias{plot_recruitment_deviations}
\title{Plot Recruitment Deviations}
\usage{
plot_recruitment_deviations(
  dat,
  geom = "point",
  make_rda = FALSE,
  figures_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame or names list of data frames returned from
\link[asar]{convert_output}. The first data frame in the list is used in
calculation of a reference line if one is present}

\item{geom}{A string stating the geom used for the plot. Default is "line".
Options include "line", "point", or "area"}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{figures_dir}{The location of the folder containing the generated figure
rda files ("figures") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}

\item{...}{Arguments called from ggplot2::geom_line or ggplotr2::geom_point}
}
\value{
Plot recruitment deviations relative to one over time from an
assessment model output(s) file translated to a standardized output. There are
options to return a `ggplot2` object or export an .rda object containing
associated caption and alternative text for the figure.
}
\description{
Plot Recruitment Deviations
}
\examples{
\dontrun{
plot_recruitment_deviations(
  dat = list("model1"=dat1,"model2"=dat2),
  geom = "point",
  make_rda = TRUE,
  figures_dir = getwd(),
  size = 3
)
}
}
