% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawn_recruitment.R
\name{plot_spawn_recruitment}
\alias{plot_spawn_recruitment}
\title{Plot Spawn-Recruit Curve}
\usage{
plot_spawn_recruitment(
  dat,
  spawning_biomass_label = "mt",
  recruitment_label = "mt",
  end_year = NULL,
  make_rda = FALSE,
  figures_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from \link[asar]{convert_output}}

\item{spawning_biomass_label}{Units for spawning biomass}

\item{recruitment_label}{units for recruitment}

\item{end_year}{last year of assessment}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{figures_dir}{The location of the folder containing the generated figure
rda files ("figures") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Plot spawning recruitment relationship from a standardized output file originating from \link[asar]{convert_output}
}
\description{
Plot Spawn-Recruit Curve
}
\examples{
\dontrun{
plot_spawn_recruitment(dat)

plot_spawn_recruitment(
  dat,
  spawning_biomass_label = "sb label",
  recruitment_label = "rec label",
  end_year = 2024,
  make_rda = TRUE,
  figures_dir = getwd()
)
}
}
