% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawning_biomass.R
\name{plot_spawning_biomass}
\alias{plot_spawning_biomass}
\title{Plot spawning biomass (SB)}
\usage{
plot_spawning_biomass(
  dat,
  geom = "line",
  group = NULL,
  facet = NULL,
  ref_line = "msy",
  unit_label = "metric tons",
  era = "time",
  module = NULL,
  scale_amount = 1,
  relative = FALSE,
  make_rda = FALSE,
  figures_dir = getwd(),
  interactive = TRUE,
  ...
)
}
\arguments{
\item{dat}{A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}. 

If inputting a list of tibbles, the first tibble's reference point defined 
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{geom}{A string stating the geom used for the plot. 

Default: "line".

Options: "line", "point", or "area"}

\item{group}{A string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.). 

Set group = "none" to summarize data over all indexing values.

Default: NULL}

\item{facet}{A string or vector of strings of a column name.

Default: NULL}

\item{ref_line}{A string specifying a reference point name.

Default: "target"

Options: (including, but not limited to) "target", "msy", and "unfished"
If the reference point is not found in the data, set ref_line = c("{name}" = value).}

\item{unit_label}{A string specifying spawning biomass unit.

Default: "metric tons"}

\item{era}{A string naming the era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{module}{(Optional) A string indicating the module_name found in `dat`.

Default: NULL

If the interactive and >1 module_name is found, user will select the 
module_name in the console. @seealso [filter_data()]}

\item{scale_amount}{A number to scale the y-axis values.

Default: 1}

\item{relative}{A logical value specifying to set y-axis values relative to 
the ref_line value.

Default: `FALSE`}

\item{make_rda}{A logical value indicating whether to save the object and 
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{A string indicating a path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}

\item{interactive}{A logical value indicating if the environment is interactive.

Default: `FALSE`}

\item{...}{Arguments called from \link[ggplot2]{geom_line} or \link[ggplot2]{geom_point}}
}
\value{
Plot spawning biomass over time from the results of an assessment model translated to
the a standardized output (\link[stockplotr]{convert_output}). There are options to return a
\link[ggplot2]{ggplot} object or export an rda object containing associated
caption and alternative text for the figure.
}
\description{
Plot spawning biomass with a reference line as a dashed line. The figure can
also be made relative to this reference line rather than in absolute units.
}
\note{
All plotting functions automatically recognize indexing variables and will 
use them in groupings and/or facetting. @seealso [process_data()].
}
\examples{
plot_spawning_biomass(
  dat = stockplotr:::example_data,
  geom = "line",
  ref_line = "msy",
  unit_label = "mt",
  scale_amount = 1000,
  interactive = FALSE,
  module = "TIME_SERIES",
  linewidth = 1.5
)
plot_spawning_biomass(
  dat = stockplotr:::example_data,
  relative = TRUE,
  ref_line = "msy",
  module = "TIME_SERIES"
)
plot_spawning_biomass(
 dat = stockplotr:::example_data,
 ref_line = c("target" = 10),
 interactive = FALSE,
 module = "TIME_SERIES"
)
}
\seealso{
[plot_timeseries()], [calculate_reference_point()],
[reference_line()], [filter_data()], [process_data()]
}
