% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawning_biomass.R
\name{plot_spawning_biomass}
\alias{plot_spawning_biomass}
\title{Plot spawning biomass (SB)}
\usage{
plot_spawning_biomass(
  dat,
  geom = "line",
  group = NULL,
  facet = NULL,
  ref_line = "msy",
  unit_label = "metric tons",
  scale_amount = 1,
  relative = FALSE,
  make_rda = FALSE,
  figures_dir = getwd(),
  ...
)
}
\arguments{
\item{dat}{A data frame or names list of data frames returned from
\link[asar]{convert_output}. The first data frame in the list is used in
calculation of a reference line if one is present}

\item{geom}{A string stating the geom used for the plot. Default is "line".
Options include "line", "point", or "area"}

\item{group}{a string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.). Currently can only have one level of grouping.}

\item{facet}{a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.)}

\item{ref_line}{A string specifying the type of reference you want to
compare spawning biomass to. The default is `"msy"`, which looks for
`"spawning_biomass_target"` in the `"label"` column of `dat`. The actual
searching in `dat` is case agnostic and will work with either upper- or
lower-case letters but you must use one of the options specified in the
default list to ensure that the label on the figure looks correct
regardless of how it is specified in `dat`. Other possibilities may include
"target", "MSY", and "unfished".}

\item{unit_label}{units for spawning_biomass}

\item{scale_amount}{A number describing how much to scale down the quantities
shown on the y axis. For example, scale_amount = 100 would scale down a value
from 500,000 --> 5,000. This scale will be reflected in the y axis label.}

\item{relative}{A logical value specifying if the resulting figures should
be relative spawning biomass. The default is `FALSE`. `ref_line` indicates
which reference point to use.}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{figures_dir}{The location of the folder containing the generated figure
rda files ("figures") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}

\item{...}{Arguments called from ggplot2::geom_line or ggplotr2::geom_point}
}
\value{
Plot spawning biomass from the results of an assessment model translated to
the standard output. The [ggplot2::ggplot()] object is returned for further
modifications if needed.
}
\description{
Plot spawning biomass with a reference line as a dashed line. The figure can
also be made relative to this reference line rather than in absolute units.
}
\examples{
\dontrun{
plot_spawning_biomass(
  dat = list('base_model'=dat1,'sensitivity1'=dat2,'sensitivity2'=dat3),
  geom = "line",
  group= "sex",
  facet = "fleet",
  ref_line = "target",
  unit_label = "mt",
  scale_amount = 1000
)
}
}
