% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock_recruitment.R
\name{plot_stock_recruitment}
\alias{plot_stock_recruitment}
\title{Plot Stock Recruit Relationship}
\usage{
plot_stock_recruitment(
  dat,
  spawning_biomass_label = "mt",
  recruitment_label = "mt",
  interactive = TRUE,
  module = NULL,
  scale_amount = 1,
  make_rda = FALSE,
  figures_dir = getwd()
)
}
\arguments{
\item{dat}{A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}. 

If inputting a list of tibbles, the first tibble's reference point defined 
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{spawning_biomass_label}{Units for spawning biomass}

\item{recruitment_label}{units for recruitment}

\item{interactive}{Indicate whether the environment the
plot is being made in is interactive. By default, this
is set to false. If true, dependent on your data, a
option menu will pop-up.}

\item{module}{(Optional) A string indicating the module_name found in `dat`.

Default: NULL

If the interactive and >1 module_name is found, user will select the 
module_name in the console. @seealso [filter_data()]}

\item{scale_amount}{A number to scale the y-axis values.

Default: 1}

\item{make_rda}{A logical value indicating whether to save the object and 
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{figures_dir}{A string indicating a path to the "figures" folder.

Default: `getwd()`

The folder is created within the path if it does not exist.}
}
\value{
Plot the stock recruitment relationship from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}). There are options to return a
[ggplot2::ggplot()] object or export an rda object containing associated
caption and alternative text for the figure.
}
\description{
Plot Stock Recruit Relationship
}
\examples{
plot_stock_recruitment(
  dat = stockplotr:::example_data,
  interactive = FALSE,
  spawning_biomass_label = "metric tons",
  recruitment_label = "metric tons",
  module = "SPAWN_RECRUIT"
)
}
