% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series trends}
\usage{
plot_timeseries(
  dat,
  x,
  y,
  geom,
  xlab = "Year",
  ylab = NULL,
  group = NULL,
  facet = NULL,
  ...
)
}
\arguments{
\item{geom}{type of geom to use for plotting found in ggplot2 (e.g. "point", "line", etc.)}

\item{xlab}{a string of the x-axis label (default is "Year")}

\item{ylab}{a string of the y-axis label. If NULL, it will be set to the name of `y`.}

\item{group}{a string of a single column that groups the data (e.g. "fleet", "sex", "area", etc.)}

\item{facet}{a string of a single column that facets the data (e.g. "year", "area", etc.)}

\item{...}{}
}
\description{
Plot time series trends
}
