% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series trends}
\usage{
plot_timeseries(
  dat,
  x = "year",
  y = "estimate",
  geom = "line",
  xlab = "Year",
  ylab = NULL,
  group = NULL,
  facet = NULL,
  ...
)

plot_timeseries(
  dat,
  x = "year",
  y = "estimate",
  geom = "line",
  xlab = "Year",
  ylab = NULL,
  group = NULL,
  facet = NULL,
  ...
)
}
\arguments{
\item{dat}{filtered data frame from standard output file(s) preformatted for
the target label from \link[stockplotr]{filter_data}}

\item{x}{a string of the column name of data used to plot on the x-axis (default
is "year")}

\item{y}{a string of the column name of data used to plot on the y-axis (default
is "estimate")}

\item{geom}{type of geom to use for plotting found in ggplot2 (e.g. "point",
"line", etc.). Default is "line". Other options are "point" and "area".}

\item{xlab}{a string of the x-axis label (default is "Year")}

\item{ylab}{a string of the y-axis label. If NULL, it will be set to the name
of `y`.}

\item{group}{a string of a single column that groups the data (e.g. "fleet",
"sex", "area", etc.). Currently can only have one level of grouping.}

\item{facet}{a string or vector of strings of a column that facets the data
(e.g. "year", "area", etc.)}

\item{...}{inherited arguments from internal functions from ggplot2::geom_xx}

\item{hline}{indicate true or false to place a horizontal line at 1}
}
\value{
Create a time series plot for a stock assessment report. The user
has options to create a line, point, or area plot where the x-axis is year
and Y can vary for any time series quantity. Currently, grouping is
restricted to one group where faceting can be any number of facets.

<<<<<<< HEAD
Create a time series plot for a stock assessment report. The user
has options to create a line, point, or area plot where the x-axis is year
=======
Create a time series plot for a stock assessment report. The user 
has options to create a line, point, or area plot where the x-axis is year 
>>>>>>> 3b39d6c (Update documentation)
and Y can vary for any time series quantity. Currently, grouping is
restricted to one group where faceting can be any number of facets.
}
\description{
Plot time series trends

Create plot with error
}
\examples{
\dontrun{
plot_timeseries(dat,
  x = "year",
  y = "estimate",
  geom = "line",
  xlab = "Year",
  ylab = "Biomass",
  group = "fleet",
  facet = "area"
)
}
\dontrun{
plot_timeseries(dat,
               x = "year",
               y = "estimate",
               geom = "line",
               xlab = "Year",
               ylab = "Biomass",
               group = "fleet",
               facet = "area")
}
}
