% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Post processing of filtered data}
\usage{
process_data(dat, group = NULL, facet = NULL, method = "sum")
}
\arguments{
\item{dat}{Pre-filtered data from \link[stockplotr]{filter_data} following a
long format data.}

\item{group}{A string identifying the indexing variable of the data. If you
want to just summarize the data across all factors, set group = "none".}

\item{facet}{A string or vector of strings identifying the faceting
variable(s) of the data.}

\item{method}{A string describing the method of summarizing data when group
is set to "none". Options are "sum" or "mean". Default is "sum".}
}
\value{
Automatically detects potential grouping and faceting data from a 
dataframe output from \link[stockplotr]{filter_data}. It returns the following 
list of 3 objects:
\item{data}{A data frame of the processed data ready for plotting.}
\item{group}{A string identifying the grouping variable of the data.
If NULL, no grouping variable is identified. If not NULL, the function
will verify that the data is indexed by this variables, otherwise it will 
overwrite it to a different, valid indexed variable or NULL.}
\item{facet}{A string or vector of strings identifying the faceting
variable(s) of the data. If NULL, no faceting variable is identified. Any
identified indexed variables found in this function will be added to facet.}
}
\description{
Post processing of filtered data
}
\examples{
{
filtered <- stockplotr::filter_data(
 dat = stockplotr:::example_data,
 label_name = "fishing_mortality$",
 geom = "line",
 module = "TIME_SERIES"
)
process_data(dat = filtered, method = "sum")
}
}
