% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_table}
\alias{process_table}
\title{Processing for tables}
\usage{
process_table(dat, group = NULL, method = "sum", label = NULL)
}
\arguments{
\item{dat}{Pre-filtered data from \link[stockplotr]{filter_data} following a
long format data.}

\item{group}{A string identifying the indexing variable of the data. If you
want to just summarize the data across all factors, set group = "none".}

\item{method}{A string describing the method of summarizing data when group
is set to "none". Options are "sum" or "mean". Default is "sum".}

\item{label}{A string or vector of strings identifying the label values to filter the data.}
}
\value{
A dataframe of processed data ready for formatting into a table. Input is an object created with \link[stockplotr]{filter_data}.
}
\description{
Processing for tables
}
\examples{
{
  filtered <- filter_data(
    dat = stockplotr:::example_data,
    label_name = "landings",
    geom = "line",
    era = "time"
  )
  process_table(dat = filtered, method = "sum")
}
}
