% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{reference_line}
\alias{reference_line}
\title{Preformatted reference line}
\usage{
reference_line(
  plot,
  dat,
  era = "time",
  label_name,
  reference,
  relative = FALSE,
  scale_amount = 1
)
}
\arguments{
\item{plot}{a ggplot2 object where the reference line will be added}

\item{dat}{standard data frame where reference point should be extracted}

\item{era}{A string naming the era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{label_name}{string of the name of the quantity that users want to
extract the reference point from}

\item{reference}{string. name of the reference point such as "msy",
"unfished", or "target"}

\item{relative}{A logical value specifying to set y-axis values relative to 
the ref_line value.

Default: `FALSE`}

\item{scale_amount}{A number to scale the y-axis values.

Default: 1}
}
\value{
a ggplot2 geom_hline object for a reference point that can be added
to a plot
}
\description{
Preformatted reference line
}
\examples{
\dontrun{
reference_line(dat, "biomass", "msy")
}
}
