% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{reference_line}
\alias{reference_line}
\title{Pre-formatted reference line}
\usage{
reference_line(dat, label_name, reference)
}
\arguments{
\item{dat}{standard data frame where reference point should be extracted}

\item{label_name}{string of the name of the quantity that users want to 
extract the reference point from}

\item{reference}{string. name of the reference point such as "msy", 
"unfished", or "target"}
}
\value{
a ggplot2 geom_hline object for a reference point that can be added 
to a plot
}
\description{
Pre-formatted reference line
}
\examples{
\dontrun{
reference_line(dat, "biomass", "msy")
}
}
