% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{reference_line}
\alias{reference_line}
\title{Preformatted reference line}
\usage{
reference_line(
  plot,
  dat,
  label_name,
  reference,
  relative = FALSE,
  scale_amount = 1
)
}
\arguments{
\item{plot}{a ggplot2 object where the reference line will be added}

\item{dat}{standard data frame where reference point should be extracted}

\item{label_name}{string of the name of the quantity that users want to 
extract the reference point from}

\item{reference}{string. name of the reference point such as "msy", 
"unfished", or "target"}

\item{relative}{A logical value specifying if the resulting figures should
be relative spawning biomass. The default is `FALSE`. `ref_line` indicates
which reference point to use.}

\item{scale_amount}{A number describing how much to scale down the quantities
shown on the y axis. For example, scale_amount = 100 would scale down a value
from 500,000 --> 5,000. This scale will be reflected in the y axis label.}
}
\value{
a ggplot2 geom_hline object for a reference point that can be added 
to a plot
}
\description{
Preformatted reference line
}
\examples{
\dontrun{
reference_line(dat, "biomass", "msy")
}
}
