% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_biomass_abundance_catch.R
\name{table_biomass_abundance_catch}
\alias{table_biomass_abundance_catch}
\title{Biomass, abundance, and catch time series table}
\usage{
table_biomass_abundance_catch(
  dat,
  unit_label = c(biomass = "mt", abundance = "fish", catch = "mt"),
  era = NULL,
  interactive = TRUE,
  group = NULL,
  method = "sum",
  module = NULL,
  label = NULL,
  make_rda = FALSE,
  tables_dir = getwd()
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{Character vector. Abbreviated unit label for each quantity

Default: c("biomass" = "mt", "abundance" = "fish", "catch" = "mt")}

\item{era}{String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{group}{A string of a single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{method}{A string describing the method of summarizing data when group
is set to "none".

Default: "sum"

Options: "sum" or "mean"}

\item{module}{Character vector. (Optional) Module name found in `dat$module_name`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{label}{The label that will be chosen from the input file. If unspecified,
the function will search the "label" column and use the first matching label
in this ordered list: "landings_weight",  "landings_numbers", "landings_expected",
"landings_predicted", "landings".

Default: NULL}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{tables_dir}{The location of the folder containing the table
rda files ("tables") that will be created if the argument `make_rda` = TRUE.
Default: the working directory (`getwd()`)}
}
\value{
A table of biomass, abundance, catch, and spawning biomass.
}
\description{
Biomass, abundance, and catch time series table
}
\details{
There are
options to return a [gt::gt()] object or export an rda object containing
a gt-based table, caption, and LaTeX-based table.
}
\examples{
table_biomass_abundance_catch(stockplotr::example_data)

table_biomass_abundance_catch(stockplotr::example_data,
  unit_label = c("biomass" = "kg", "abundance" = "eggs", "catch" = "kg"),
  module = c("TIME_SERIES", "CATCH", "TIME_SERIES"),
  interactive = FALSE
)
}
\seealso{
[convert_output()], [filter_data()], [process_table()], [export_kqs()], [insert_kqs()], [create_rda()]
}
