% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bnc.R
\name{table_bnc}
\alias{table_bnc}
\title{Biomass, abundance, and catch time series table}
\usage{
table_bnc(
  dat,
  end_year = NULL,
  biomass_unit_label = "mt",
  catch_unit_label = "mt",
  sb_unit_label = "mt",
  make_rda = FALSE,
  tables_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from \link[asar]{convert_output}}

\item{end_year}{last year of assessment}

\item{biomass_unit_label}{abbreviated units for biomass}

\item{catch_unit_label}{abbreviated units for catch}

\item{sb_unit_label}{abbreviated units for spawning biomass}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}
}
\value{
Create a table of biomass, abundance, catch, and spawning biomass through all years of
the assessment model output translated to a standard structure.There are
options to return a {flextable} object or export an rda object containing
associated caption for the table.
}
\description{
Biomass, abundance, and catch time series table
}
\examples{
\dontrun{
table_bnc(dat)

table_bnc(dat,
  end_year = 2024,
  biomass_unit_label = "b label",
  catch_unit_label = "catch label",
  make_rda = TRUE,
  tables_dir = getwd()
)
}
}
