% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_catch.R
\name{table_catch}
\alias{table_catch}
\title{Catch (table) by fleet, gear type, sector}
\usage{
table_catch(
  dat,
  unit_label = "mt",
  era = NULL,
  interactive = TRUE,
  module = NULL,
  scale_amount = 1,
  digits = 2,
  make_rda = FALSE,
  tables_dir = getwd()
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{String. Abbreviated catch units

Default: "mt"}

\item{era}{String. Era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{module}{Character vector. (Optional) Module name found in `dat$module_name`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{scale_amount}{Number. A number to scale the y-axis values.

Default: 1}

\item{digits}{Number. Numeric value indicating the number of digits catch values in the
table will be rounded to.

Default: 2}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{tables_dir}{Path. The location of the folder containing the generated table
rda files ("tables") that will be created if the argument `make_rda` = TRUE.

Default: the working directory (`getwd()`)}
}
\value{
A table ready of landed catch by fleet and year.
}
\description{
Catch (table) by fleet, gear type, sector
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a [gt::gt()] object or export an rda object
containing a gt-based table, caption, and LaTeX-based table.
}
\examples{
table_catch(stockplotr::example_data)

table_catch(
  stockplotr::example_data,
  unit_label = "lbs"
)
}
\seealso{
[convert_output()], [filter_data()], [process_table()], [export_kqs()], [insert_kqs()], [create_rda()]
}
