% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_index.R
\name{table_index}
\alias{table_index}
\title{Index of abundance table}
\usage{
table_index(
  dat,
  era = NULL,
  interactive = TRUE,
  group = NULL,
  method = "sum",
  module = NULL,
  label = NULL,
  digits = 2,
  make_rda = FALSE,
  tables_dir = getwd()
)
}
\arguments{
\item{dat}{A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{era}{A string naming the era of data.

Default: "time"

Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{interactive}{A logical value indicating if the environment is interactive.

Default: `FALSE`}

\item{group}{A string of a single column that groups the data.

Set group = "none" to summarize data over all indexing values.

Default: NULL
Options: Including, but not limited to: "year", "area", "fleet", "sex", "none", NULL}

\item{method}{A string describing the method of summarizing data when group
is set to "none".

Default: "sum"

Options: "sum" or "mean"}

\item{module}{(Optional) A string indicating the module_name found in `dat`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{label}{The label that will be chosen from the input file. If unspecified,
the function will search the "label" column and use the first matching label
in this ordered list: "index_weight",  "index_numbers", "index_expected",
"index_predicted", "index".

Default: NULL}

\item{digits}{Numeric value indicating the number of digits values in the
table will be rounded to.

Default: 2}

\item{make_rda}{A logical value indicating whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{tables_dir}{The location of the folder containing the generated table
rda files ("tables") that will be created if the argument `make_rda` = TRUE.

Default: the working directory (`getwd()`)}
}
\value{
A table of observed annual index of abundance plus error,
stratified by fleet.
}
\description{
Index of abundance table
}
\details{
The input is from an assessment model output file
translated to a standardized output (\link[stockplotr]{convert_output}).
There are options to return a [gt::gt()] object or export an rda object
containing a gt-based table, caption, and LaTeX-based table.
}
\examples{
\dontrun{
table_index(dat)

table_index(
  dat,
  make_rda = TRUE,
  tables_dir = getwd()
)
}
}
\seealso{
[convert_output()], [filter_data()], [process_table()], [export_kqs()], [insert_kqs()], [create_rda()]
}
