% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_landings.R
\name{table_landings}
\alias{table_landings}
\title{Landed catch by fleet and year table}
\usage{
table_landings(
  dat,
  unit_label = "mt",
  era = NULL,
  interactive = TRUE,
  group = NULL,
  method = "sum",
  module = NULL,
  label = NULL,
  make_rda = FALSE,
  tables_dir = getwd()
)
}
\arguments{
\item{dat}{A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}. 
If inputting a list of tibbles, the first tibble's reference point defined 
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{Abbreviated units of landings}

\item{era}{A string naming the era of data.
Default: "time"
Options: "early", "time", "fore" (forecast), or NULL (all data)}

\item{interactive}{A logical value indicating if the environment is interactive.
Default: `FALSE`}

\item{group}{A string identifying the indexing variable of the data. If you
want to just summarize the data across all factors, set group = "none".}

\item{method}{A string describing the method of summarizing data when group
is set to "none". Options are "sum" or "mean". Default is "sum".}

\item{module}{(Optional) A string indicating the module_name found in `dat`.
Default: NULL
If the interactive and >1 module_name is found, user will select the 
module_name in the console. @seealso [filter_data()]}

\item{label}{The label that will be chosen from the input file. If unspecified,
the function will search the "label" column and use the first matching label
in this ordered list: "landings_weight",  "landings_numbers", "landings_expected",
"landings_predicted", "landings".}

\item{make_rda}{A logical value indicating whether to save the object and 
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default: `FALSE`.}

\item{tables_dir}{The location of the folder containing the generated table
rda files ("tables") that will be created if the argument `make_rda` = TRUE.}
}
\value{
Create a table ready for a stock assessment report of landed catch by
fleet and year.
}
\description{
Landed catch by fleet and year table
}
\examples{
table_landings(stockplotr::example_data)

table_landings(
  stockplotr::example_data,
  unit_label = "landings label",
  group = "fleet"
)
}
