% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_landings.R
\name{table_landings}
\alias{table_landings}
\title{Landed catch by fleet and year table}
\usage{
table_landings(
  dat,
  unit_label = "mt",
  era = "time",
  interactive = TRUE,
  module = NULL,
  scale_amount = 1,
  label = "landings_weight",
  make_rda = FALSE,
  tables_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame or names list of data frames (input as `list()`)
returned from \link[asar]{convert_output}. The first data frame in the list 
is used in calculation of a reference line if one is present}

\item{unit_label}{Abbreviated units of landings}

\item{era}{a string naming the era of data such as historical ("early"), current ("time"), or 
projected ("fore") data if filtering should occur. Default is set to "time" which is 
the current time. To plot all data, set era to NULL.}

\item{interactive}{TRUE/FALSE; indicate whether the environment in which the
function is operating  is interactive. This bypasses some options for
filtering when preparing data for the plot. Default is FALSE.}

\item{module}{(Optional) A string indicating the linked module_name associated
with the label for the plot if known. Default is NULL. By default, the function
will select the most relevant module if more than 1 exists.}

\item{scale_amount}{A number describing how much to scale down the quantities
shown on the y axis. For example, scale_amount = 100 would scale down a value
from 500,000 --> 5,000. This scale will be reflected in the y axis label.}

\item{label}{The label that will be chosen from the input file. If unspecified, the function will search the "label" column and use the first matching label in this ordered list: "landings_weight",  "landings_numbers", "landings_expected", "landings_predicted", "landings".}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{tables_dir}{The location of the folder containing the generated table
rda files ("tables") that will be created if the argument `make_rda` = TRUE.}
}
\value{
Create a table ready for a stock assessment report of landed catch by
fleet and year.
}
\description{
Landed catch by fleet and year table
}
\examples{
\dontrun{
table_landings(dat)

table_landings(
  dat,
  unit_label = "landings label",
  end_year = 2024,
  make_rda = TRUE,
  tables_dir = getwd()
)
}
}
