% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_landings.R
\name{table_landings}
\alias{table_landings}
\title{Landed catch by fleet and year table}
\usage{
table_landings(
  dat,
  unit_label = "mt",
  end_year = format(Sys.Date(), "\%Y"),
  make_rda = FALSE,
  tables_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame or names list of data frames returned from
\link[asar]{convert_output}. The first data frame in the list is used in
calculation of a reference line if one is present}

\item{unit_label}{Abbreviated units of landings}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".
Default is FALSE.}

\item{tables_dir}{The location of the folder containing the generated table
rda files ("tables") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Create a table ready for a stock assessment report of landed catch by
fleet and year.
}
\description{
Landed catch by fleet and year table
}
\examples{
\dontrun{
table_landings(dat)

table_landings(
  dat,
  unit_label = "landings label",
  end_year = 2024,
  make_rda = TRUE,
  tables_dir = getwd()
)
}
}
