% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_projections.R
\name{table_projections}
\alias{table_projections}
\title{Create Projection Summary Table}
\usage{
table_projections(
  dat,
  unit_label = c(catch = "mt", spawning_biomass = "mt", fishing_mortality = ""),
  interactive = TRUE,
  module = NULL,
  make_rda = FALSE,
  tables_dir = getwd()
)
}
\arguments{
\item{dat}{Data frame or list. A tibble or named list of tibbles (input as `list()`)
returned from \link[stockplotr]{convert_output}.

If inputting a list of tibbles, the first tibble's reference point defined
in `ref_line` is used to plot a reference line or calculate relative spawning biomass.}

\item{unit_label}{Character vector. Abbreviated unit label for each quantity

Default: c("catch" = "mt", "spawning_biomass" = "mt",
"fishing_mortality" = "")}

\item{interactive}{Logical. TRUE/FALSE; indicate whether the environment is interactive.

Default: `FALSE`}

\item{module}{Character vector. (Optional) Module name found in `dat$module_name`.
If selecting >1 module, place them in a vector like c("module1", "module2").

Default: NULL

If the interactive and >1 module_name is found, user will select the
module_name in the console. @seealso [filter_data()]}

\item{make_rda}{Logical. TRUE/FALSE; indicate whether to save the object and
make an automated caption and alternative text in the form of an `rda` object. If TRUE,
the rda will be exported to the folder indicated in the argument "figures_dir".

Default: `FALSE`.}

\item{tables_dir}{Path. The location of the folder containing the generated table
rda files ("tables") that will be created if the argument `make_rda` = TRUE.

Default: the working directory (`getwd()`)}
}
\value{
A formatted gt table object.
}
\description{
Generates a table showing assumed and projected years of Catch, SB, and F.
}
\details{
We would like to thank Dan Hennen for sharing his projections table
function code, which served as the foundation for this function.
}
\examples{
table_projections(
  dat = stockplotr::example_data,
  interactive = FALSE,
  module = "DERIVED_QUANTITIES"
)

}
