% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_noaa.R
\name{theme_noaa}
\alias{theme_noaa}
\title{Add NOAA theming to ggplot2 object}
\usage{
theme_noaa(...)
}
\value{
theme for NOAA Fisheries plots
}
\description{
Add NOAA theming to ggplot2 object
}
\examples{
ggplot2::ggplot() + ggplot2::geom_line(ggplot2::aes(x = 1:10, y = 1:10)) + theme_noaa()
}
