% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_table.R
\name{theme_table}
\alias{theme_table}
\title{Add NOAA formatting to a table}
\usage{
theme_table(x)
}
\arguments{
\item{x}{Object. Table from gt, flextable, or kableExtra}
}
\value{
Add the standard formatting for stock assessment reports for any table.
}
\description{
Add NOAA formatting to a table
}
\details{
Currently, the function can format table objects from:
flextable ({flextable} package, when installed), gt ({gt}), and
kable ({kableExtra}).
}
\examples{
theme_table(cars |> gt::gt())
theme_table(cars |> kableExtra::kable())
}
